improvement(workflows): unify run cancellation
