+++ apps/docs/openapi-v2-workflows.json
@@ -9986,7 +9986,7 @@
           },
           "durablyRecorded": {
             "type": "boolean",
-            "description": "Whether this request durably recorded a cancellation. Always false for a run that was already terminal, where the request is satisfied but nothing was written."
+            "description": "Whether this request durably recorded a cancellation. False when an already-cancelled run needed no further durable write."
           },
           "locallyAborted": {
             "type": "boolean",
@@ -9997,7 +9997,7 @@
             "description": "Whether a paused execution was cancelled."
           },
           "reason": {
-            "description": "Machine-readable cancellation outcome, present on every cancellation including full successes. `recorded` is the success value. `already_cancelled`, `already_completed`, and `already_failed` mean the run had already reached that terminal state, so nothing was cancelled and `durablyRecorded` is false. `redis_unavailable` and `redis_write_failed` mean the distributed cancellation signal was not written, so an already-running execution may not observe the cancellation. `paused_event_publish_failed` and `paused_database_cancel_failed` name the failing step for a paused run.",
+            "description": "Machine-readable cancellation outcome. `recorded`, `queue_cancelled`, and `already_cancelled` are successful outcomes; the remaining values identify a degraded or incomplete cancellation step.",
             "type": "string",
             "enum": [
               "recorded",
@@ -10007,7 +10007,10 @@
               "redis_unavailable",
               "redis_write_failed",
               "paused_event_publish_failed",
-              "paused_database_cancel_failed"
+              "paused_database_cancel_failed",
+              "queue_cancelled",
+              "active_resume_signal_failed",
+              "cancellation_not_finalized"
             ]
           }
         },
@@ -10021,7 +10024,7 @@
         ],
         "additionalProperties": false,
         "title": "Cancel workflow run result",
-        "description": "Outcome of a workflow run cancellation request. Cancellation is best-effort: a run already in a terminal state succeeds with no effect, reported as `durablyRecorded: false` with an `already_*` reason naming the state observed."
+        "description": "Outcome of the shared workflow-run cancellation lifecycle used by internal UI, Copilot, and v2 callers."
       },
       "CancelWorkflowRunResponse": {
         "type": "object",
