+++ apps/sim/app/api/workflows/[id]/executions/[executionId]/cancel/route.test.ts
@@ -1,1585 +1,96 @@
 /**
  * @vitest-environment node
  */
-
-import {
-  databaseMock,
-  dbChainMockFns,
-  hybridAuthMockFns,
-  posthogServerMock,
-  resetDbChainMock,
-  workflowAuthzMockFns,
-  workflowsUtilsMock,
-} from '@sim/testing'
 import { NextRequest } from 'next/server'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
+import { OrchestrationError } from '@/lib/core/orchestration/types'
 
-const {
-  mockMarkExecutionCancelled,
-  mockClearExecutionCancellation,
-  mockAbortManualExecution,
-  mockBeginPausedCancellation,
-  mockStagePausedCancellation,
-  mockBlockQueuedResumesForCancellation,
-  mockClearPausedCancellationIntent,
-  mockCompletePausedCancellation,
-  mockGetPausedCancellationStatus,
-  mockGetActiveResumeCancellationTarget,
-  mockRollbackActiveResumeCancellation,
-  mockFinalizeExecutionStream,
-  mockReadExecutionMetaState,
-  mockWriteEvent,
-  mockWriteTerminalEvent,
-  mockCancelByExecution,
-  mockGetJobQueue,
-  mockReleaseExecutionSlot,
-  mockCancelWorkflowGroupExecution,
-  mockPublishWorkflowGroupCancellationEvent,
-} = vi.hoisted(() => ({
-  mockMarkExecutionCancelled: vi.fn(),
-  mockClearExecutionCancellation: vi.fn(),
-  mockAbortManualExecution: vi.fn(),
-  mockBeginPausedCancellation: vi.fn(),
-  mockStagePausedCancellation: vi.fn(),
-  mockBlockQueuedResumesForCancellation: vi.fn(),
-  mockClearPausedCancellationIntent: vi.fn(),
-  mockCompletePausedCancellation: vi.fn(),
-  mockGetPausedCancellationStatus: vi.fn(),
-  mockGetActiveResumeCancellationTarget: vi.fn(),
-  mockRollbackActiveResumeCancellation: vi.fn(),
-  mockFinalizeExecutionStream: vi.fn(),
-  mockReadExecutionMetaState: vi.fn(),
-  mockWriteEvent: vi.fn(),
-  mockWriteTerminalEvent: vi.fn(),
-  mockCancelByExecution: vi.fn(),
-  mockGetJobQueue: vi.fn(),
-  mockReleaseExecutionSlot: vi.fn(),
-  mockCancelWorkflowGroupExecution: vi.fn(),
-  mockPublishWorkflowGroupCancellationEvent: vi.fn(),
+const mocks = vi.hoisted(() => ({
+  cancel: vi.fn(),
+  getSession: vi.fn(),
 }))
 
-vi.mock('@/lib/core/async-jobs', () => ({
-  getJobQueue: mockGetJobQueue,
+vi.mock('@/lib/auth', () => ({ getSession: mocks.getSession }))
+
+vi.mock('@/lib/workflows/application/cancel-run', () => ({
+  cancelWorkflowRun: { operation: { id: 'workflows.runs.cancel' }, execute: mocks.cancel },
 }))
 
-vi.mock('@/lib/billing/calculations/usage-reservation', () => ({
-  releaseExecutionSlot: mockReleaseExecutionSlot,
-}))
+import { POST } from '@/app/api/workflows/[id]/executions/[executionId]/cancel/route'
 
-vi.mock('@/lib/execution/cancellation', () => ({
-  markExecutionCancelled: (...args: unknown[]) => mockMarkExecutionCancelled(...args),
-  clearExecutionCancellation: (...args: unknown[]) => mockClearExecutionCancellation(...args),
-}))
+const principal = { kind: 'session' as const, userId: 'user-1', sessionId: 'session-1' }
+const context = { params: Promise.resolve({ id: 'workflow-1', executionId: 'execution-1' }) }
 
-vi.mock('@/lib/execution/manual-cancellation', () => ({
-  abortManualExecution: (...args: unknown[]) => mockAbortManualExecution(...args),
-}))
-
-vi.mock('@/lib/workflows/executor/human-in-the-loop-manager', () => ({
-  PauseResumeManager: {
-    beginPausedCancellation: (...args: unknown[]) => mockBeginPausedCancellation(...args),
-    stagePausedCancellation: (...args: unknown[]) => mockStagePausedCancellation(...args),
-    blockQueuedResumesForCancellation: (...args: unknown[]) =>
-      mockBlockQueuedResumesForCancellation(...args),
-    clearPausedCancellationIntent: (...args: unknown[]) =>
-      mockClearPausedCancellationIntent(...args),
-    completePausedCancellation: (...args: unknown[]) => mockCompletePausedCancellation(...args),
-    getPausedCancellationStatus: (...args: unknown[]) => mockGetPausedCancellationStatus(...args),
-    getActiveResumeCancellationTarget: (...args: unknown[]) =>
-      mockGetActiveResumeCancellationTarget(...args),
-    rollbackActiveResumeCancellation: (...args: unknown[]) =>
-      mockRollbackActiveResumeCancellation(...args),
-  },
-}))
-
-vi.mock('@/lib/workflows/utils', () => workflowsUtilsMock)
-
-vi.mock('@/lib/posthog/server', () => posthogServerMock)
-
-vi.mock('@/lib/table/workflow-group-cancellation', () => ({
-  cancelWorkflowGroupExecution: (...args: unknown[]) => mockCancelWorkflowGroupExecution(...args),
-  publishWorkflowGroupCancellationEvent: (...args: unknown[]) =>
-    mockPublishWorkflowGroupCancellationEvent(...args),
-}))
-
-vi.mock('@/lib/execution/event-buffer', () => ({
-  finalizeExecutionStream: (...args: unknown[]) => mockFinalizeExecutionStream(...args),
-  readExecutionMetaState: (...args: unknown[]) => mockReadExecutionMetaState(...args),
-  createExecutionEventWriter: () => ({
-    write: (...args: unknown[]) => mockWriteEvent(...args),
-    writeTerminal: (...args: unknown[]) => mockWriteTerminalEvent(...args),
-    close: vi.fn().mockResolvedValue(undefined),
-  }),
-}))
-
-import { cancelWorkflowExecutionContract } from '@/lib/api/contracts/workflows'
-import { cancelWorkflowExecutionPostAuth } from '@/lib/execution/cancel-workflow-execution-post-auth'
-import { POST as cancelExecution } from './route'
-
-/**
- * Drives the route and validates every success body against the contract that
- * `requestJson` enforces on the client. The route builds its responses by hand
- * rather than through a declarative builder, so nothing else checks that the
- * two agree — and a `reason` the contract omits makes the client throw on a
- * cancellation that actually applied. Routing every case in this suite through
- * here covers each `NextResponse.json` shape the route can return.
- */
-const POST = async (...args: Parameters<typeof cancelExecution>) => {
-  const response = await cancelExecution(...args)
-  if (response.status < 400) {
-    cancelWorkflowExecutionContract.response.schema.parse(await response.clone().json())
-  }
-  return response
-}
-
-const makeRequest = () =>
-  new NextRequest('http://localhost/api/workflows/wf-1/executions/ex-1/cancel', {
-    method: 'POST',
-  })
-
-const makeParams = () => ({ params: Promise.resolve({ id: 'wf-1', executionId: 'ex-1' }) })
-
-const ACTIVE_RESUME_TARGET = {
-  resumeEntryId: 'resume-entry-1',
-  pausedExecutionId: 'paused-1',
-  parentExecutionId: 'ex-1',
-  resumeExecutionId: 'resume-ex-1',
+function request() {
+  return new NextRequest(
+    'http://localhost/api/workflows/workflow-1/executions/execution-1/cancel',
+    { method: 'POST' }
+  )
 }
 
 describe('POST /api/workflows/[id]/executions/[executionId]/cancel', () => {
   beforeEach(() => {
     vi.clearAllMocks()
-    resetDbChainMock()
-    hybridAuthMockFns.mockCheckHybridAuth.mockResolvedValue({ success: true, userId: 'user-1' })
-    workflowAuthzMockFns.mockAuthorizeWorkflowByWorkspacePermission.mockResolvedValue({
-      allowed: true,
-      workflow: { workspaceId: 'workspace-1' },
+    mocks.getSession.mockResolvedValue({
+      user: { id: principal.userId },
+      session: { id: principal.sessionId },
     })
-    dbChainMockFns.limit.mockResolvedValue([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: null,
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    dbChainMockFns.returning.mockResolvedValue([{ status: 'cancelled' }])
-    mockCancelByExecution.mockReset().mockResolvedValue(0)
-    mockGetJobQueue.mockReset().mockResolvedValue({ cancelByExecution: mockCancelByExecution })
-    mockReleaseExecutionSlot.mockReset().mockResolvedValue(undefined)
-    mockCancelWorkflowGroupExecution.mockReset().mockResolvedValue({ kind: 'not_workflow_group' })
-    mockPublishWorkflowGroupCancellationEvent.mockReset().mockResolvedValue(undefined)
-    mockClearExecutionCancellation.mockReset().mockResolvedValue(undefined)
-    mockMarkExecutionCancelled
-      .mockReset()
-      .mockResolvedValue({ durablyRecorded: false, reason: 'redis_unavailable' })
-    mockAbortManualExecution.mockReset().mockReturnValue(false)
-    mockBeginPausedCancellation.mockReset().mockResolvedValue(false)
-    mockStagePausedCancellation.mockReset().mockResolvedValue({ kind: 'not_paused' })
-    mockBlockQueuedResumesForCancellation.mockReset().mockResolvedValue(false)
-    mockClearPausedCancellationIntent.mockReset().mockResolvedValue(undefined)
-    mockCompletePausedCancellation.mockReset().mockResolvedValue(false)
-    mockGetPausedCancellationStatus.mockReset().mockResolvedValue(null)
-    mockGetActiveResumeCancellationTarget.mockReset().mockResolvedValue(null)
-    mockRollbackActiveResumeCancellation.mockReset().mockResolvedValue(true)
-    mockFinalizeExecutionStream.mockReset().mockResolvedValue(true)
-    mockReadExecutionMetaState.mockReset().mockResolvedValue({ status: 'missing' })
-    mockWriteEvent.mockReset().mockResolvedValue({ eventId: 1 })
-    mockWriteTerminalEvent.mockReset().mockResolvedValue({ eventId: 1 })
-  })
-
-  it('returns success when cancellation was durably recorded', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: true,
-      reason: 'recorded',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toEqual({
+    mocks.cancel.mockResolvedValue({
       success: true,
-      executionId: 'ex-1',
+      executionId: 'execution-1',
       redisAvailable: true,
       durablyRecorded: true,
       locallyAborted: false,
       pausedCancelled: false,
       reason: 'recorded',
-    })
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      {
-        workflowId: 'wf-1',
-        executionId: 'ex-1',
-      },
-      'standalone'
-    )
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1', {
-      executionDeadlineAt: null,
-    })
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
-  })
-
-  it('atomically claims one workflow-group attempt before signalling it', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: true,
-      reason: 'recorded',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      executionId: 'ex-1',
-      redisAvailable: true,
-      durablyRecorded: true,
-      reason: 'recorded',
-    })
-    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledWith({
+      workflowId: 'workflow-1',
       workspaceId: 'workspace-1',
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-    })
-    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
-      mockMarkExecutionCancelled.mock.invocationCallOrder[0]
-    )
-    expect(mockMarkExecutionCancelled.mock.invocationCallOrder[0]).toBeLessThan(
-      mockPublishWorkflowGroupCancellationEvent.mock.invocationCallOrder[0]
-    )
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-    expect(mockStagePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
-    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
-  })
-
-  it('keeps a claimed group cancellation retryable when signalling fails', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: false,
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledOnce()
-    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
-    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
-    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-  })
-
-  it('accepts an exact in-process group abort without cancelling its carrier', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-    mockAbortManualExecution.mockReturnValue(true)
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      durablyRecorded: false,
-      locallyAborted: true,
-      reason: 'redis_unavailable',
-    })
-    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-  })
-
-  it('does not cancel a shared carrier when an idle workflow-group pause appears', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockStagePausedCancellation
-      .mockResolvedValueOnce({ kind: 'not_paused' })
-      .mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      pausedCancelled: true,
-      reason: 'recorded',
-    })
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-    expect(mockStagePausedCancellation.mock.invocationCallOrder[1]).toBeLessThan(
-      mockWriteTerminalEvent.mock.invocationCallOrder[0]
-    )
-    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
-  })
-
-  it('leaves a late workflow-group pause retryable when event publication fails', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockStagePausedCancellation
-      .mockResolvedValueOnce({ kind: 'not_paused' })
-      .mockResolvedValue({ kind: 'idle' })
-    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: false,
-      pausedCancelled: false,
-      reason: 'paused_event_publish_failed',
-    })
-    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
-    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
-      mockWriteTerminalEvent.mock.invocationCallOrder[0]
-    )
-    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
-    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
-  })
-
-  it('cancels an active workflow-group resume without cancelling its shared carrier', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockStagePausedCancellation.mockResolvedValue({
-      kind: 'active_resume',
-      target: ACTIVE_RESUME_TARGET,
-    })
-    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-    mockCancelByExecution.mockResolvedValue(1)
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      pausedCancelled: true,
-      reason: 'recorded',
-    })
-    expect(mockCancelByExecution).toHaveBeenCalledOnce()
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      { workflowId: 'wf-1', executionId: 'ex-1' },
-      'resume'
-    )
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('resume-ex-1', {
-      executionDeadlineAt: null,
-    })
-    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
-    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
-      mockMarkExecutionCancelled.mock.invocationCallOrder[0]
-    )
-    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
-    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
-      mockCompletePausedCancellation.mock.invocationCallOrder[0]
-    )
-  })
-
-  it('keeps a claimed active group resume retryable when no stop backend accepts it', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockStagePausedCancellation.mockResolvedValue({
-      kind: 'active_resume',
-      target: ACTIVE_RESUME_TARGET,
-    })
-    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: false,
-      durablyRecorded: true,
-      locallyAborted: false,
-      pausedCancelled: false,
-      reason: 'active_resume_signal_failed',
-    })
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('resume-ex-1', {
-      executionDeadlineAt: null,
-    })
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      { workflowId: 'wf-1', executionId: 'ex-1' },
-      'resume'
-    )
-    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
-      mockMarkExecutionCancelled.mock.invocationCallOrder[0]
-    )
-    expect(mockRollbackActiveResumeCancellation).not.toHaveBeenCalled()
-    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
-  })
-
-  it('rechecks for a pause after the group terminal claim waits on persistence', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockStagePausedCancellation
-      .mockResolvedValueOnce({ kind: 'not_paused' })
-      .mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      pausedCancelled: true,
-      reason: 'recorded',
-    })
-    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
-      mockStagePausedCancellation.mock.invocationCallOrder[1]
-    )
-    expect(mockStagePausedCancellation.mock.invocationCallOrder[1]).toBeLessThan(
-      mockWriteTerminalEvent.mock.invocationCallOrder[0]
-    )
-    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
-      mockCompletePausedCancellation.mock.invocationCallOrder[0]
-    )
-    expect(mockWriteTerminalEvent).toHaveBeenCalledOnce()
-  })
-
-  it('rechecks for a regular pause after the terminal claim waits on persistence', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockStagePausedCancellation
-      .mockResolvedValueOnce({ kind: 'not_paused' })
-      .mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      pausedCancelled: true,
-      reason: 'recorded',
-    })
-    expect(databaseMock.db.update.mock.invocationCallOrder[0]).toBeLessThan(
-      mockStagePausedCancellation.mock.invocationCallOrder[1]
-    )
-    expect(mockStagePausedCancellation.mock.invocationCallOrder[1]).toBeLessThan(
-      mockWriteTerminalEvent.mock.invocationCallOrder[0]
-    )
-    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
-      mockCompletePausedCancellation.mock.invocationCallOrder[0]
-    )
-  })
-
-  it('uses generic cancellation for a regular Table-trigger-block execution', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: null,
-        status: 'running',
-        trigger: 'table',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: true,
-      reason: 'recorded',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    expect(mockCancelWorkflowGroupExecution).not.toHaveBeenCalled()
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1', {
-      executionDeadlineAt: null,
-    })
-    expect(mockCancelByExecution).toHaveBeenCalledOnce()
-  })
-
-  it('returns 409 without generic cancellation when the exact group attempt is stale', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'conflict',
-      status: 'no_longer_active',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(409)
-    await expect(response.json()).resolves.toEqual({
-      error: 'Workflow group execution cannot be cancelled while no_longer_active',
-    })
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-    expect(mockStagePausedCancellation).toHaveBeenCalledOnce()
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
-    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
-  })
-
-  it('terminalizes a durable workflow-group log after its table sidecar was deleted', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({ kind: 'cancelled_without_sidecar' })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      executionId: 'ex-1',
-      reason: 'recorded',
-    })
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledOnce()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
-    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
-  })
-
-  it('returns unsuccessful response when Redis is unavailable', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toEqual({
-      success: false,
-      executionId: 'ex-1',
-      redisAvailable: false,
-      durablyRecorded: false,
-      locallyAborted: false,
-      pausedCancelled: false,
-      reason: 'redis_unavailable',
     })
   })
 
-  it('returns unsuccessful response when Redis persistence fails', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_write_failed',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toEqual({
-      success: false,
-      executionId: 'ex-1',
-      redisAvailable: true,
-      durablyRecorded: false,
-      locallyAborted: false,
-      pausedCancelled: false,
-      reason: 'redis_write_failed',
-    })
-  })
-
-  it('returns success when local fallback aborts execution without Redis durability', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-    mockAbortManualExecution.mockReturnValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
+  it('passes the authenticated principal and canonical run input to the application use case', async () => {
+    const response = await POST(request(), context)
 
     expect(response.status).toBe(200)
     await expect(response.json()).resolves.toEqual({
       success: true,
-      executionId: 'ex-1',
-      redisAvailable: false,
-      durablyRecorded: false,
-      locallyAborted: true,
-      pausedCancelled: false,
-      reason: 'redis_unavailable',
-    })
-  })
-
-  it('returns success when the queue backend cancels the active job', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-    mockCancelByExecution.mockResolvedValue(1)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      executionId: 'ex-1',
-      durablyRecorded: false,
-      locallyAborted: false,
-      reason: 'queue_cancelled',
-    })
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
-  })
-
-  it('cancels a queued execution before its workflow log exists', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([])
-    mockCancelByExecution.mockResolvedValueOnce(1)
-    mockMarkExecutionCancelled.mockResolvedValueOnce({
-      durablyRecorded: true,
-      reason: 'recorded',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toEqual({
-      success: true,
-      executionId: 'ex-1',
+      executionId: 'execution-1',
       redisAvailable: true,
       durablyRecorded: true,
       locallyAborted: false,
       pausedCancelled: false,
-      reason: 'queue_cancelled',
-    })
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      {
-        workflowId: 'wf-1',
-        executionId: 'ex-1',
-      },
-      'standalone'
-    )
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1')
-    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('ex-1')
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
-  })
-
-  it('does not use an unscoped local abort before its workflow log exists', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([])
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(404)
-    await expect(response.json()).resolves.toEqual({ error: 'Execution not found' })
-    expect(mockAbortManualExecution).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      {
-        workflowId: 'wf-1',
-        executionId: 'ex-1',
-      },
-      'standalone'
-    )
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
-  })
-
-  it('returns success when a paused HITL execution is cancelled directly in the database', async () => {
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toEqual({
-      success: true,
-      executionId: 'ex-1',
-      redisAvailable: true,
-      durablyRecorded: true,
-      locallyAborted: false,
-      pausedCancelled: true,
       reason: 'recorded',
     })
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockWriteTerminalEvent).toHaveBeenCalledWith(
-      expect.objectContaining({
-        type: 'execution:cancelled',
-        executionId: 'ex-1',
-        workflowId: 'wf-1',
-      }),
-      'cancelled'
-    )
-    expect(mockFinalizeExecutionStream).not.toHaveBeenCalled()
-  })
-
-  it('claims the paused workflow-group sidecar before publishing and finalizing', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
+    expect(mocks.cancel).toHaveBeenCalledWith({
+      principal,
+      input: {
+        workflowId: 'workflow-1',
+        runId: 'execution-1',
+        abortSignal: expect.any(AbortSignal),
       },
-    ])
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
+      request: expect.any(NextRequest),
     })
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      durablyRecorded: true,
-      pausedCancelled: true,
-      reason: 'recorded',
-    })
-    expect(mockStagePausedCancellation.mock.invocationCallOrder[0]).toBeLessThan(
-      mockWriteTerminalEvent.mock.invocationCallOrder[0]
-    )
-    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
-      mockWriteTerminalEvent.mock.invocationCallOrder[0]
-    )
-    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
-      mockCompletePausedCancellation.mock.invocationCallOrder[0]
-    )
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-    expect(mockWriteTerminalEvent).toHaveBeenCalledOnce()
-    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
   })
 
-  it('keeps a paused workflow-group cancellation reserved when event publication fails', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'running',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
+  it('rejects unauthenticated requests before invoking cancellation', async () => {
+    mocks.getSession.mockResolvedValue(null)
 
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: false,
-      pausedCancelled: false,
-      reason: 'paused_event_publish_failed',
-    })
-    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
-    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
-    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-  })
-
-  it('publishes paused cancellation event even when Redis cancellation is recorded', async () => {
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      executionId: 'ex-1',
-      durablyRecorded: true,
-      pausedCancelled: true,
-    })
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockWriteTerminalEvent).toHaveBeenCalledWith(
-      expect.objectContaining({
-        type: 'execution:cancelled',
-        executionId: 'ex-1',
-        workflowId: 'wf-1',
-      }),
-      'cancelled'
-    )
-    expect(mockFinalizeExecutionStream).not.toHaveBeenCalled()
-  })
-
-  it('does not confirm paused cancellation when terminal event publication fails', async () => {
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toEqual({
-      success: false,
-      executionId: 'ex-1',
-      redisAvailable: false,
-      durablyRecorded: true,
-      locallyAborted: false,
-      pausedCancelled: false,
-      reason: 'paused_event_publish_failed',
-    })
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
-    expect(mockWriteTerminalEvent).toHaveBeenCalledWith(
-      expect.objectContaining({
-        type: 'execution:cancelled',
-        executionId: 'ex-1',
-        workflowId: 'wf-1',
-      }),
-      'cancelled'
-    )
-    expect(mockFinalizeExecutionStream).not.toHaveBeenCalled()
-  })
-
-  it('finishes reconciliation when the pause row is already cancelled', async () => {
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      pausedCancelled: true,
-      reason: 'recorded',
-    })
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
-      mockCompletePausedCancellation.mock.invocationCallOrder[0]
-    )
-  })
-
-  it('returns 401 when auth fails', async () => {
-    hybridAuthMockFns.mockCheckHybridAuth.mockResolvedValue({
-      success: false,
-      error: 'Unauthorized',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
+    const response = await POST(request(), context)
 
     expect(response.status).toBe(401)
+    expect(mocks.cancel).not.toHaveBeenCalled()
   })
 
-  it('returns 403 when workflow access is denied', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    workflowAuthzMockFns.mockAuthorizeWorkflowByWorkspacePermission.mockResolvedValue({
-      allowed: false,
-      message: 'Access denied',
-      status: 403,
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(403)
-  })
-
-  it('conceals a workflow outside the asserted Copilot workspace', async () => {
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      assertedWorkspaceId: 'workspace-other',
-    })
-
-    expect(response.status).toBe(404)
-    await expect(response.json()).resolves.toEqual({ error: 'Execution not found' })
-    expect(databaseMock.db.select).not.toHaveBeenCalled()
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-  })
-
-  it('stops before lookup when cancellation is aborted during authorization', async () => {
-    const controller = new AbortController()
-    workflowAuthzMockFns.mockAuthorizeWorkflowByWorkspacePermission.mockImplementationOnce(
-      async () => {
-        controller.abort()
-        return { allowed: true, workflow: { workspaceId: 'workspace-1' } }
-      }
+  it('projects application conflicts without reimplementing cancellation errors', async () => {
+    mocks.cancel.mockRejectedValue(
+      new OrchestrationError('conflict', 'Execution cannot be cancelled while completed')
     )
 
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      abortSignal: controller.signal,
-    })
+    const response = await POST(request(), context)
 
     expect(response.status).toBe(409)
-    await expect(response.json()).resolves.toEqual({
-      error: 'Request aborted before workflow run cancellation could be applied.',
-    })
-    expect(databaseMock.db.select).not.toHaveBeenCalled()
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-  })
-
-  it('stops before mutation when cancellation is aborted during execution lookup', async () => {
-    const controller = new AbortController()
-    dbChainMockFns.limit.mockImplementationOnce(async () => {
-      controller.abort()
-      return [
-        {
-          executionDeadlineAt: null,
-          executionOrigin: null,
-          status: 'running',
-          workspaceId: 'workspace-1',
-        },
-      ]
-    })
-
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      abortSignal: controller.signal,
-    })
-
-    expect(response.status).toBe(409)
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockAbortManualExecution).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-  })
-
-  it('rolls back pause staging when cancellation is aborted during staging', async () => {
-    const controller = new AbortController()
-    mockStagePausedCancellation.mockImplementationOnce(async () => {
-      controller.abort()
-      return { kind: 'idle' }
-    })
-
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      abortSignal: controller.signal,
-    })
-
-    expect(response.status).toBe(409)
-    expect(mockClearPausedCancellationIntent).toHaveBeenCalledWith('ex-1', 'wf-1')
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockAbortManualExecution).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-  })
-
-  it('rolls back an active resume staged while cancellation is aborted', async () => {
-    const controller = new AbortController()
-    mockStagePausedCancellation.mockImplementationOnce(async () => {
-      controller.abort()
-      return { kind: 'active_resume', target: ACTIVE_RESUME_TARGET }
-    })
-
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      abortSignal: controller.signal,
-    })
-
-    expect(response.status).toBe(409)
-    expect(mockRollbackActiveResumeCancellation).toHaveBeenCalledWith(
-      'ex-1',
-      'wf-1',
-      'resume-entry-1'
-    )
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-  })
-
-  it('rolls back pause staging for an already-cancelled execution when cancellation is aborted', async () => {
-    const controller = new AbortController()
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: null,
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockStagePausedCancellation.mockImplementationOnce(async () => {
-      controller.abort()
-      return { kind: 'idle' }
-    })
-
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      abortSignal: controller.signal,
-    })
-
-    expect(response.status).toBe(409)
-    expect(mockClearPausedCancellationIntent).toHaveBeenCalledWith('ex-1', 'wf-1')
-    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
-    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
-    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-  })
-
-  it('finishes cancellation when an aborted active-resume stage cannot be rolled back', async () => {
-    const controller = new AbortController()
-    mockStagePausedCancellation.mockImplementationOnce(async () => {
-      controller.abort()
-      return { kind: 'active_resume', target: ACTIVE_RESUME_TARGET }
-    })
-    mockRollbackActiveResumeCancellation.mockResolvedValueOnce(false)
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      abortSignal: controller.signal,
-    })
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({ success: true, pausedCancelled: true })
-    expect(mockRollbackActiveResumeCancellation).toHaveBeenCalledWith(
-      'ex-1',
-      'wf-1',
-      'resume-entry-1'
-    )
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('resume-ex-1', {
-      executionDeadlineAt: null,
-    })
-  })
-
-  it('returns 404 when the execution does not belong to the workflow', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([])
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(404)
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      {
-        workflowId: 'wf-1',
-        executionId: 'ex-1',
-      },
-      'standalone'
-    )
-  })
-
-  it('treats an already-cancelled execution as an idempotent success', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      { executionDeadlineAt: null, status: 'cancelled', workspaceId: 'workspace-1' },
-    ])
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
     await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      durablyRecorded: false,
-      reason: 'already_cancelled',
-    })
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('ex-1')
-  })
-
-  it('reconciles the exact sidecar when a workflow-group log is already cancelled', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'already_cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      reason: 'already_cancelled',
-    })
-    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledWith({
-      workspaceId: 'workspace-1',
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-    })
-    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1', {
-      executionDeadlineAt: null,
-    })
-    expect(mockPublishWorkflowGroupCancellationEvent).toHaveBeenCalledOnce()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-  })
-
-  it('finishes workflow-group reconciliation when abort arrives during its durable commit', async () => {
-    const controller = new AbortController()
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockCancelWorkflowGroupExecution.mockImplementationOnce(async () => {
-      controller.abort()
-      return {
-        kind: 'already_cancelled',
-        tableId: 'table-1',
-        rowId: 'row-1',
-        groupId: 'group-1',
-      }
-    })
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await cancelWorkflowExecutionPostAuth({
-      workflowId: 'wf-1',
-      executionId: 'ex-1',
-      userId: 'user-1',
-      abortSignal: controller.signal,
-    })
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({ success: true, pausedCancelled: true })
-    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
-    expect(mockPublishWorkflowGroupCancellationEvent).toHaveBeenCalledOnce()
-    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
-  })
-
-  it('does not finalize an already-cancelled group retry until exact stop is accepted', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'already_cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: false,
-      redisAvailable: false,
-      reason: 'redis_unavailable',
-    })
-    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
-    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
-    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-  })
-
-  it('repairs a stranded active resume when the group log is already cancelled', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockStagePausedCancellation.mockResolvedValue({
-      kind: 'active_resume',
-      target: ACTIVE_RESUME_TARGET,
-    })
-    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
-    mockAbortManualExecution.mockReturnValue(true)
-    mockCancelByExecution.mockResolvedValue(1)
-    mockCompletePausedCancellation.mockResolvedValue(true)
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'already_cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      locallyAborted: true,
-      pausedCancelled: true,
-      reason: 'already_cancelled',
-    })
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      { workflowId: 'wf-1', executionId: 'ex-1' },
-      'resume'
-    )
-    expect(mockAbortManualExecution).toHaveBeenCalledWith('resume-ex-1')
-    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
-    expect(mockWriteTerminalEvent).toHaveBeenCalledOnce()
-    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
-      mockCompletePausedCancellation.mock.invocationCallOrder[0]
-    )
-  })
-
-  it('repairs a stranded active resume when a regular workflow log is already cancelled', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: null,
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockStagePausedCancellation.mockResolvedValue({
-      kind: 'active_resume',
-      target: ACTIVE_RESUME_TARGET,
-    })
-    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
-    mockAbortManualExecution.mockReturnValue(true)
-    mockCancelByExecution.mockResolvedValue(1)
-    mockCompletePausedCancellation.mockResolvedValue(true)
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: true,
-      locallyAborted: true,
-      pausedCancelled: true,
-      reason: 'already_cancelled',
-    })
-    expect(mockCancelByExecution).toHaveBeenCalledWith(
-      { workflowId: 'wf-1', executionId: 'ex-1' },
-      'resume'
-    )
-    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
-      mockCompletePausedCancellation.mock.invocationCallOrder[0]
-    )
-    expect(mockCancelWorkflowGroupExecution).not.toHaveBeenCalled()
-  })
-
-  it('keeps an already-cancelled active resume retryable when publication fails', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: null,
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockStagePausedCancellation.mockResolvedValue({
-      kind: 'active_resume',
-      target: ACTIVE_RESUME_TARGET,
-    })
-    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({
-      success: false,
-      pausedCancelled: false,
-      reason: 'paused_event_publish_failed',
-    })
-    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
-  })
-
-  it('fails closed when an already-cancelled group sidecar cannot be reconciled', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'conflict',
-      status: 'completed',
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(409)
-    await expect(response.json()).resolves.toEqual({
-      error: 'Workflow group execution cannot be reconciled while completed',
-    })
-  })
-
-  it('reports a reconciliation failure for an already-cancelled group sidecar', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      {
-        executionDeadlineAt: null,
-        executionOrigin: 'workflow_group',
-        status: 'cancelled',
-        workspaceId: 'workspace-1',
-      },
-    ])
-    mockCancelWorkflowGroupExecution.mockRejectedValue(new Error('database unavailable'))
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(500)
-    await expect(response.json()).resolves.toEqual({ error: 'database unavailable' })
-  })
-
-  it('returns 409 when the execution is already terminal', async () => {
-    dbChainMockFns.limit.mockResolvedValueOnce([
-      { status: 'completed', workspaceId: 'workspace-1' },
-    ])
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(409)
-    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
-    expect(mockCancelByExecution).not.toHaveBeenCalled()
-  })
-
-  it('returns 409 when completion wins the terminal database race', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    const returning = vi.fn().mockResolvedValue([])
-    const where = vi.fn(() => ({ returning }))
-    databaseMock.db.update.mockReturnValueOnce({ set: vi.fn(() => ({ where })) })
-    dbChainMockFns.limit
-      .mockResolvedValueOnce([
-        { executionDeadlineAt: null, status: 'running', workspaceId: 'workspace-1' },
-      ])
-      .mockResolvedValueOnce([{ status: 'completed' }])
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(409)
-    await expect(response.json()).resolves.toEqual({
       error: 'Execution cannot be cancelled while completed',
     })
-    expect(returning).toHaveBeenCalledOnce()
-    expect(mockClearExecutionCancellation).toHaveBeenCalledWith('ex-1')
-    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
-    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-  })
-
-  it('rolls back paused cancellation intent when resume completion wins the log claim', async () => {
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-    const returning = vi.fn().mockResolvedValue([])
-    const where = vi.fn(() => ({ returning }))
-    databaseMock.db.update.mockReturnValueOnce({ set: vi.fn(() => ({ where })) })
-    dbChainMockFns.limit
-      .mockResolvedValueOnce([
-        { executionDeadlineAt: null, status: 'running', workspaceId: 'workspace-1' },
-      ])
-      .mockResolvedValueOnce([{ status: 'completed' }])
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(409)
-    await expect(response.json()).resolves.toEqual({
-      error: 'Execution cannot be cancelled while completed',
-    })
-    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
-    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
-    expect(mockClearPausedCancellationIntent).toHaveBeenCalledWith('ex-1', 'wf-1')
-  })
-
-  it('treats a concurrent cancellation as an idempotent success', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    dbChainMockFns.limit
-      .mockResolvedValueOnce([
-        { executionDeadlineAt: null, status: 'running', workspaceId: 'workspace-1' },
-      ])
-      .mockResolvedValueOnce([{ status: 'cancelled' }])
-    dbChainMockFns.returning.mockResolvedValueOnce([])
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    await expect(response.json()).resolves.toMatchObject({ success: true, reason: 'recorded' })
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
-    expect(mockWriteTerminalEvent).toHaveBeenCalledTimes(1)
-    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('ex-1')
-  })
-
-  it('updates execution log status in DB when durably recorded', async () => {
-    const mockWhere = vi.fn().mockResolvedValue(undefined)
-    const mockSet = vi.fn(() => ({ where: mockWhere }))
-    databaseMock.db.update.mockReturnValueOnce({ set: mockSet })
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: true,
-      reason: 'recorded',
-    })
-
-    await POST(makeRequest(), makeParams())
-
-    expect(databaseMock.db.update).toHaveBeenCalled()
-    expect(mockSet).toHaveBeenCalledWith({
-      status: 'cancelled',
-      endedAt: expect.any(Date),
-      totalDurationMs: expect.anything(),
-      executionDeadlineAt: null,
-    })
-  })
-
-  it('updates execution log status in DB when locally aborted', async () => {
-    const mockWhere = vi.fn().mockResolvedValue(undefined)
-    const mockSet = vi.fn(() => ({ where: mockWhere }))
-    databaseMock.db.update.mockReturnValueOnce({ set: mockSet })
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
-    })
-    mockAbortManualExecution.mockReturnValue(true)
-
-    await POST(makeRequest(), makeParams())
-
-    expect(databaseMock.db.update).toHaveBeenCalled()
-    expect(mockSet).toHaveBeenCalledWith({
-      status: 'cancelled',
-      endedAt: expect.any(Date),
-      totalDurationMs: expect.anything(),
-      executionDeadlineAt: null,
-    })
-  })
-
-  it('claims the execution log before finalizing a paused cancellation', async () => {
-    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
-
-    await POST(makeRequest(), makeParams())
-
-    expect(databaseMock.db.update).toHaveBeenCalled()
-  })
-
-  it('does not confirm cancellation until the terminal database update succeeds', async () => {
-    mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: true,
-      reason: 'recorded',
-    })
-    databaseMock.db.update.mockReturnValueOnce({
-      set: vi.fn(() => ({
-        where: vi.fn(() => {
-          throw new Error('DB connection failed')
-        }),
-      })),
-    })
-
-    const response = await POST(makeRequest(), makeParams())
-
-    expect(response.status).toBe(200)
-    const data = await response.json()
-    expect(data).toMatchObject({
-      success: false,
-      reason: 'cancellation_not_finalized',
-    })
-    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
   })
 })
