+++ apps/sim/lib/api/contracts/v2/workflows.ts
@@ -1865,26 +1865,21 @@ export const v2CancelWorkflowRunDataSchema = z
     durablyRecorded: z
       .boolean()
       .describe(
-        'Whether this request durably recorded a cancellation. Always false for a run that was already terminal, where the request is satisfied but nothing was written.'
+        'Whether this request durably recorded a cancellation. False when an already-cancelled run needed no further durable write.'
       ),
     locallyAborted: z.boolean().describe('Whether an in-process execution was aborted.'),
     pausedCancelled: z.boolean().describe('Whether a paused execution was cancelled.'),
-    /**
-     * Always emitted by the cancellation service — it is not a partial-failure
-     * marker. `recorded` is the full-success value; the `already_*` values name
-     * a terminal no-op; the rest name the step that degraded.
-     */
     reason: cancelWorkflowExecutionReasonSchema
       .optional()
       .describe(
-        'Machine-readable cancellation outcome, present on every cancellation including full successes. `recorded` is the success value. `already_cancelled`, `already_completed`, and `already_failed` mean the run had already reached that terminal state, so nothing was cancelled and `durablyRecorded` is false. `redis_unavailable` and `redis_write_failed` mean the distributed cancellation signal was not written, so an already-running execution may not observe the cancellation. `paused_event_publish_failed` and `paused_database_cancel_failed` name the failing step for a paused run.'
+        'Machine-readable cancellation outcome. `recorded`, `queue_cancelled`, and `already_cancelled` are successful outcomes; the remaining values identify a degraded or incomplete cancellation step.'
       ),
   })
   .meta({
     id: 'CancelWorkflowRunResult',
     title: 'Cancel workflow run result',
     description:
-      'Outcome of a workflow run cancellation request. Cancellation is best-effort: a run already in a terminal state succeeds with no effect, reported as `durablyRecorded: false` with an `already_*` reason naming the state observed.',
+      'Outcome of the shared workflow-run cancellation lifecycle used by internal UI, Copilot, and v2 callers.',
   })
 export type V2CancelWorkflowRunData = z.output<typeof v2CancelWorkflowRunDataSchema>
