+++ apps/sim/lib/copilot/tools/handlers/workflow/mutations.ts
@@ -1,5 +1,4 @@
 import { createLogger } from '@sim/logger'
-import { cancelWorkflowExecutionContract } from '@/lib/api/contracts/workflows'
 import { createCopilotWorkspaceApiKey } from '@/lib/api-key/application/create-api-key'
 import { messageForCopilotApplicationError } from '@/lib/copilot/application/error'
 import { executeCopilotApiKeyUseCase } from '@/lib/copilot/application/execute-api-key-use-case'
@@ -7,7 +6,6 @@ import {
   executeCopilotWorkflowUseCase,
   messageForCopilotWorkflowError,
 } from '@/lib/copilot/application/execute-workflow-use-case'
-import { requireTrustedCopilotExecutionContext } from '@/lib/copilot/auth/application-delegation'
 import type { ExecutionContext, ToolCallResult } from '@/lib/copilot/request/types'
 import {
   TOOL_EFFECT_PHASE,
@@ -17,7 +15,7 @@ import {
 import { requireCopilotWorkspace } from '@/lib/copilot/tools/server/workspace-scope'
 import { decodeVfsPathSegments, encodeVfsPathSegments } from '@/lib/copilot/vfs/path-utils'
 import { PlatformEvents } from '@/lib/core/telemetry'
-import { cancelWorkflowExecutionPostAuth } from '@/lib/execution/cancel-workflow-execution-post-auth'
+import { cancelWorkflowRun } from '@/lib/workflows/application/cancel-run'
 import { createWorkflow } from '@/lib/workflows/application/create-workflow'
 import { moveWorkflowsBulk } from '@/lib/workflows/application/move-workflows-bulk'
 import {
@@ -153,12 +151,6 @@ function resolveRunTriggerBlockId(params: { triggerBlockId?: unknown }): string 
 
 function resolveInputFromExecutionId(value: unknown): string | undefined {
   return typeof value === 'string' && value.trim().length > 0 ? value.trim() : undefined
-}
-
-function readCancellationRouteError(body: unknown): string | undefined {
-  if (!body || typeof body !== 'object' || !('error' in body)) return undefined
-  const error = body.error
-  return typeof error === 'string' && error.trim() ? error : undefined
 }
 
 function copilotRunLifecycle(context: ExecutionContext) {
@@ -310,24 +302,12 @@ export async function executeCancelWorkflowRun(
       context,
       'Request aborted before workflow run cancellation could be applied.'
     )
-    const trustedContext = requireTrustedCopilotExecutionContext(context)
-    const response = await cancelWorkflowExecutionPostAuth({
+    const result = await executeCopilotWorkflowUseCase(context, cancelWorkflowRun, {
       workflowId,
-      executionId,
-      userId: trustedContext.userId,
-      assertedWorkspaceId: trustedContext.workspaceId,
+      runId: executionId,
+      assertedWorkspaceId: context.workspaceId,
       ...(context.abortSignal ? { abortSignal: context.abortSignal } : {}),
     })
-    const responseBody: unknown = await response.json()
-    if (!response.ok) {
-      return {
-        success: false,
-        error:
-          readCancellationRouteError(responseBody) ||
-          `Failed to cancel workflow run (${response.status})`,
-      }
-    }
-    const result = cancelWorkflowExecutionContract.response.schema.parse(responseBody)
 
     return {
       success: result.success,
