--- apps/sim/lib/execution/cancel-workflow-execution-post-auth.ts
@@ -1,1052 +1,0 @@
-import { db } from '@sim/db'
-import { workflowExecutionLogs } from '@sim/db/schema'
-import { createLogger } from '@sim/logger'
-import { authorizeWorkflowByWorkspacePermission } from '@sim/platform-authz/workflow'
-import { toError } from '@sim/utils/errors'
-import { sleep } from '@sim/utils/helpers'
-import { and, eq, inArray } from 'drizzle-orm'
-import { NextResponse } from 'next/server'
-import type { CancelWorkflowExecutionResponse } from '@/lib/api/contracts/workflows'
-import { WORKSPACE_KEY_SCOPE_DENIED } from '@/lib/api-key/policy-messages'
-import { releaseExecutionSlot } from '@/lib/billing/calculations/usage-reservation'
-import { getJobQueue } from '@/lib/core/async-jobs'
-import type { ExecutionJobCancellationScope } from '@/lib/core/async-jobs/types'
-import {
-  clearExecutionCancellation,
-  type ExecutionCancellationRecordResult,
-  markExecutionCancelled,
-} from '@/lib/execution/cancellation'
-import { createExecutionEventWriter, readExecutionMetaState } from '@/lib/execution/event-buffer'
-import { abortManualExecution } from '@/lib/execution/manual-cancellation'
-import { cancelledExecutionLogFields } from '@/lib/logs/execution/cancellation'
-import { workflowExecutionOriginSql } from '@/lib/logs/execution-origin'
-import { captureServerEvent } from '@/lib/posthog/server'
-import {
-  cancelWorkflowGroupExecution,
-  type PublishableWorkflowGroupCancellation,
-  publishWorkflowGroupCancellationEvent,
-} from '@/lib/table/workflow-group-cancellation'
-import { PauseResumeManager } from '@/lib/workflows/executor/human-in-the-loop-manager'
-
-const logger = createLogger('CancelExecutionAPI')
-const PAUSED_CANCELLATION_DB_ATTEMPTS = 3
-const PAUSED_CANCELLATION_DB_RETRY_MS = 200
-const CANCELLATION_ABORTED_MESSAGE =
-  'Request aborted before workflow run cancellation could be applied.'
-
-/**
- * Builds the single success shape this handler returns. It hand-builds its
- * responses rather than going through a declarative builder, so nothing else
- * checks that they satisfy the contract the client validates against — and
- * several outcomes the route resolves itself (a still-queued run, an
- * already-cancelled run) ride on `success: true`, where a rejected body turns a
- * cancellation that worked into a client-side failure.
- */
-function cancellationOutcome(body: CancelWorkflowExecutionResponse) {
-  return NextResponse.json(body)
-}
-
-async function cancelQueuedExecutionJobs(
-  workflowId: string,
-  executionId: string,
-  scope: ExecutionJobCancellationScope
-): Promise<number> {
-  try {
-    const queue = await getJobQueue()
-    return await queue.cancelByExecution({ workflowId, executionId }, scope)
-  } catch (error) {
-    logger.warn('Failed to cancel queued execution jobs', {
-      workflowId,
-      executionId,
-      error: toError(error).message,
-    })
-    return 0
-  }
-}
-
-function abortLocalExecution(executionId: string): boolean {
-  try {
-    return abortManualExecution(executionId)
-  } catch (error) {
-    logger.warn('Failed to abort local execution', {
-      executionId,
-      error: toError(error).message,
-    })
-    return false
-  }
-}
-
-interface ExecutionStopSignalResult {
-  cancellation: ExecutionCancellationRecordResult
-  locallyAborted: boolean
-  queueJobsCancelled: number
-  accepted: boolean
-}
-
-interface ExecutionStopSummary extends ExecutionStopSignalResult {
-  signalledExecutionIds: Set<string>
-}
-
-type ActiveResumeCancellationTarget = NonNullable<
-  Awaited<ReturnType<typeof PauseResumeManager.getActiveResumeCancellationTarget>>
->
-
-function createExecutionStopSummary(): ExecutionStopSummary {
-  return {
-    cancellation: { durablyRecorded: false, reason: 'redis_unavailable' },
-    locallyAborted: false,
-    queueJobsCancelled: 0,
-    accepted: false,
-    signalledExecutionIds: new Set<string>(),
-  }
-}
-
-function mergeExecutionStopSignal(
-  summary: ExecutionStopSummary,
-  signalExecutionId: string,
-  result: ExecutionStopSignalResult
-): void {
-  if (result.cancellation.durablyRecorded || !summary.cancellation.durablyRecorded) {
-    summary.cancellation = result.cancellation
-  }
-  summary.locallyAborted = summary.locallyAborted || result.locallyAborted
-  summary.queueJobsCancelled += result.queueJobsCancelled
-  summary.accepted = summary.accepted || result.accepted
-  summary.signalledExecutionIds.add(signalExecutionId)
-}
-
-/**
- * Commits cancellation after the caller's final abort check. Once signalling begins, the
- * operation must finish reconciliation because workers may already have observed the durable,
- * local, or queued signal; attempting to honor a later abort could revive only part of a run.
- */
-async function signalExecutionStop(args: {
-  workflowId: string
-  signalExecutionId: string
-  queueBindingExecutionId?: string
-  executionDeadlineAt: Date | null
-  queueScope?: ExecutionJobCancellationScope
-}): Promise<ExecutionStopSignalResult> {
-  const cancellation = await markExecutionCancelled(args.signalExecutionId, {
-    executionDeadlineAt: args.executionDeadlineAt,
-  })
-  const locallyAborted = abortLocalExecution(args.signalExecutionId)
-  const queueJobsCancelled = args.queueScope
-    ? await cancelQueuedExecutionJobs(
-        args.workflowId,
-        args.queueBindingExecutionId ?? args.signalExecutionId,
-        args.queueScope
-      )
-    : 0
-  return {
-    cancellation,
-    locallyAborted,
-    queueJobsCancelled,
-    accepted: cancellation.durablyRecorded || locallyAborted || queueJobsCancelled > 0,
-  }
-}
-
-async function signalAndRecordActiveResumeStop(args: {
-  workflowId: string
-  executionId: string
-  executionDeadlineAt: Date | null
-  target: ActiveResumeCancellationTarget
-  summary: ExecutionStopSummary
-}): Promise<boolean> {
-  const signal = await signalExecutionStop({
-    workflowId: args.workflowId,
-    signalExecutionId: args.target.resumeExecutionId,
-    queueBindingExecutionId: args.executionId,
-    executionDeadlineAt: args.executionDeadlineAt,
-    queueScope: 'resume',
-  })
-  mergeExecutionStopSignal(args.summary, args.target.resumeExecutionId, signal)
-  return didActiveResumeStop(args.executionId, args.workflowId, args.target, signal)
-}
-
-async function didActiveResumeStop(
-  executionId: string,
-  workflowId: string,
-  target: ActiveResumeCancellationTarget,
-  signal: ExecutionStopSignalResult
-): Promise<boolean> {
-  if (signal.accepted) return true
-  const currentTarget = await PauseResumeManager.getActiveResumeCancellationTarget(
-    executionId,
-    workflowId
-  )
-  if (currentTarget && currentTarget.resumeEntryId !== target.resumeEntryId) {
-    logger.warn('A replacement resume became active while cancellation was staged', {
-      executionId,
-      previousResumeEntryId: target.resumeEntryId,
-      currentResumeEntryId: currentTarget.resumeEntryId,
-    })
-  }
-  return currentTarget === null
-}
-
-type PausedCancellationStage = Awaited<
-  ReturnType<typeof PauseResumeManager.stagePausedCancellation>
->
-
-function isPausedCancellationStage(
-  stage: PausedCancellationStage
-): stage is Exclude<PausedCancellationStage, { kind: 'not_paused' }> {
-  return stage.kind !== 'not_paused'
-}
-
-async function clearStopSignalMarkers(summary: ExecutionStopSummary): Promise<void> {
-  await Promise.all(
-    [...summary.signalledExecutionIds].map((executionId) => clearExecutionCancellation(executionId))
-  )
-}
-
-type ExecutionLogCancellationClaim =
-  | { kind: 'cancelled' }
-  | { kind: 'conflict'; status: string }
-  | { kind: 'not_found' }
-
-async function claimExecutionLogCancellation(args: {
-  executionId: string
-  workflowId: string
-  workspaceId: string
-}): Promise<ExecutionLogCancellationClaim> {
-  const now = new Date()
-  const [cancelledExecution] = await db
-    .update(workflowExecutionLogs)
-    .set(cancelledExecutionLogFields(now))
-    .where(
-      and(
-        eq(workflowExecutionLogs.executionId, args.executionId),
-        eq(workflowExecutionLogs.workflowId, args.workflowId),
-        eq(workflowExecutionLogs.workspaceId, args.workspaceId),
-        inArray(workflowExecutionLogs.status, ['running', 'pending'])
-      )
-    )
-    .returning({ status: workflowExecutionLogs.status })
-
-  if (cancelledExecution?.status === 'cancelled') return { kind: 'cancelled' }
-
-  const currentExecution = await db
-    .select({ status: workflowExecutionLogs.status })
-    .from(workflowExecutionLogs)
-    .where(
-      and(
-        eq(workflowExecutionLogs.executionId, args.executionId),
-        eq(workflowExecutionLogs.workflowId, args.workflowId),
-        eq(workflowExecutionLogs.workspaceId, args.workspaceId)
-      )
-    )
-    .limit(1)
-    .then((rows) => rows[0])
-
-  if (!currentExecution) return { kind: 'not_found' }
-  if (currentExecution.status === 'cancelled') return { kind: 'cancelled' }
-  return { kind: 'conflict', status: currentExecution.status }
-}
-
-async function completePausedCancellationWithRetry(
-  executionId: string,
-  workflowId: string,
-  options: { logMissing?: boolean } = {}
-): Promise<boolean> {
-  for (let attempt = 1; attempt <= PAUSED_CANCELLATION_DB_ATTEMPTS; attempt++) {
-    try {
-      const cancelled = await PauseResumeManager.completePausedCancellation(executionId, workflowId)
-      if (cancelled) {
-        logger.info('Paused execution cancelled in database', { executionId, attempt })
-        return true
-      }
-      if (options.logMissing !== false) {
-        logger.warn('Paused execution cancellation could not be completed in database', {
-          executionId,
-          attempt,
-        })
-      }
-      return false
-    } catch (error) {
-      logger.warn('Failed to complete paused execution cancellation in database', {
-        executionId,
-        attempt,
-        error,
-      })
-      if (attempt < PAUSED_CANCELLATION_DB_ATTEMPTS) {
-        await sleep(PAUSED_CANCELLATION_DB_RETRY_MS)
-      }
-    }
-  }
-  return false
-}
-
-async function ensureCancellationEventPublished(
-  executionId: string,
-  workflowId: string,
-  context: { workspaceId?: string; userId?: string } = {}
-): Promise<boolean> {
-  try {
-    const metaState = await readExecutionMetaState(executionId)
-    if (metaState.status === 'found' && metaState.meta.status === 'cancelled') {
-      return true
-    }
-  } catch (error) {
-    logger.warn('Failed to read execution state before publishing cancellation', {
-      executionId,
-      error: toError(error).message,
-    })
-  }
-
-  const writer = createExecutionEventWriter(executionId, {
-    workspaceId: context.workspaceId,
-    workflowId,
-    userId: context.userId,
-  })
-  try {
-    await writer.writeTerminal(
-      {
-        type: 'execution:cancelled',
-        timestamp: new Date().toISOString(),
-        executionId,
-        workflowId,
-        data: { duration: 0 },
-      },
-      'cancelled'
-    )
-    return true
-  } catch (error) {
-    logger.warn('Failed to publish execution cancellation event', {
-      executionId,
-      error,
-    })
-    return false
-  } finally {
-    await writer.close().catch((error) => {
-      logger.warn('Failed to close cancellation event writer', {
-        executionId,
-        error,
-      })
-    })
-  }
-}
-
-export interface CancelWorkflowExecutionPostAuthInput {
-  workflowId: string
-  executionId: string
-  userId: string
-  assertedWorkspaceId?: string
-  abortSignal?: AbortSignal
-  workspaceApiKeyScope?: { workspaceId?: string }
-}
-
-function cancellationAbortedResponse(abortSignal?: AbortSignal): NextResponse | null {
-  return abortSignal?.aborted
-    ? NextResponse.json({ error: CANCELLATION_ABORTED_MESSAGE }, { status: 409 })
-    : null
-}
-
-async function rollbackPausedCancellationAfterAbort(args: {
-  stage: PausedCancellationStage
-  workflowId: string
-  executionId: string
-  abortSignal?: AbortSignal
-}): Promise<NextResponse | null> {
-  const abortResponse = cancellationAbortedResponse(args.abortSignal)
-  if (!abortResponse) return null
-
-  if (args.stage.kind === 'active_resume') {
-    const rolledBack = await PauseResumeManager.rollbackActiveResumeCancellation(
-      args.executionId,
-      args.workflowId,
-      args.stage.target.resumeEntryId
-    )
-    if (!rolledBack) {
-      logger.warn('Aborted cancellation could not be rolled back; completing cancellation', {
-        executionId: args.executionId,
-        activeResumeEntryId: args.stage.target.resumeEntryId,
-      })
-      return null
-    }
-  } else if (args.stage.kind === 'idle') {
-    await PauseResumeManager.clearPausedCancellationIntent(args.executionId, args.workflowId)
-  }
-
-  return abortResponse
-}
-
-function resolveCancellationReason(args: {
-  activeResumeSignalFailed: boolean
-  pauseReconciliationFailed: boolean
-  effectivePausedCancellationPath: boolean
-  cancellationEventPublished: boolean
-  pausedCancelled: boolean
-  stopSummary: ExecutionStopSummary
-}): NonNullable<CancelWorkflowExecutionResponse['reason']> {
-  if (args.activeResumeSignalFailed) return 'active_resume_signal_failed'
-  if (args.pauseReconciliationFailed) return 'paused_database_cancel_failed'
-  if (args.effectivePausedCancellationPath && !args.cancellationEventPublished) {
-    return 'paused_event_publish_failed'
-  }
-  if (args.effectivePausedCancellationPath && !args.pausedCancelled) {
-    return 'paused_database_cancel_failed'
-  }
-  if (args.effectivePausedCancellationPath) return 'recorded'
-  if (
-    args.stopSummary.queueJobsCancelled > 0 &&
-    !args.stopSummary.cancellation.durablyRecorded &&
-    !args.stopSummary.locallyAborted
-  ) {
-    return 'queue_cancelled'
-  }
-  return args.stopSummary.cancellation.reason
-}
-
-export async function cancelWorkflowExecutionPostAuth({
-  workflowId,
-  executionId,
-  userId,
-  assertedWorkspaceId,
-  abortSignal,
-  workspaceApiKeyScope,
-}: CancelWorkflowExecutionPostAuthInput): Promise<NextResponse> {
-  try {
-    const workflowAuthorization = await authorizeWorkflowByWorkspacePermission({
-      workflowId,
-      userId,
-      action: 'write',
-    })
-    if (!workflowAuthorization.allowed) {
-      return NextResponse.json(
-        { error: workflowAuthorization.message || 'Access denied' },
-        { status: workflowAuthorization.status }
-      )
-    }
-
-    if (
-      workspaceApiKeyScope &&
-      workflowAuthorization.workflow?.workspaceId !== workspaceApiKeyScope.workspaceId
-    ) {
-      return NextResponse.json({ error: WORKSPACE_KEY_SCOPE_DENIED }, { status: 403 })
-    }
-
-    if (
-      assertedWorkspaceId !== undefined &&
-      workflowAuthorization.workflow?.workspaceId !== assertedWorkspaceId
-    ) {
-      return NextResponse.json({ error: 'Execution not found' }, { status: 404 })
-    }
-
-    const postAuthorizationAbort = cancellationAbortedResponse(abortSignal)
-    if (postAuthorizationAbort) return postAuthorizationAbort
-
-    const execution = await db
-      .select({
-        executionDeadlineAt: workflowExecutionLogs.executionDeadlineAt,
-        executionOrigin: workflowExecutionOriginSql(),
-        status: workflowExecutionLogs.status,
-        workspaceId: workflowExecutionLogs.workspaceId,
-      })
-      .from(workflowExecutionLogs)
-      .where(
-        and(
-          eq(workflowExecutionLogs.executionId, executionId),
-          eq(workflowExecutionLogs.workflowId, workflowId)
-        )
-      )
-      .limit(1)
-      .then((rows) => rows[0])
-
-    const postLookupAbort = cancellationAbortedResponse(abortSignal)
-    if (postLookupAbort) return postLookupAbort
-
-    if (!execution) {
-      const queueJobsCancelled = await cancelQueuedExecutionJobs(
-        workflowId,
-        executionId,
-        'standalone'
-      )
-      if (queueJobsCancelled > 0) {
-        const locallyAborted = abortLocalExecution(executionId)
-        const cancellation = await markExecutionCancelled(executionId)
-        await PauseResumeManager.blockQueuedResumesForCancellation(executionId, workflowId).catch(
-          (error) => {
-            logger.warn('Failed to block queued resumes after queued-run cancellation', {
-              executionId,
-              error,
-            })
-          }
-        )
-        await releaseExecutionSlot(executionId).catch((error) => {
-          logger.warn('Failed to release reservation after queued-run cancellation', {
-            executionId,
-            error,
-          })
-        })
-
-        const workspaceId = workflowAuthorization.workflow?.workspaceId
-        captureServerEvent(
-          userId,
-          'workflow_execution_cancelled',
-          { workflow_id: workflowId, workspace_id: workspaceId ?? '' },
-          workspaceId ? { groups: { workspace: workspaceId } } : undefined
-        )
-
-        return cancellationOutcome({
-          success: true,
-          executionId,
-          redisAvailable: cancellation.reason !== 'redis_unavailable',
-          durablyRecorded: cancellation.durablyRecorded,
-          locallyAborted,
-          pausedCancelled: false,
-          reason: 'queue_cancelled',
-        })
-      }
-
-      return NextResponse.json({ error: 'Execution not found' }, { status: 404 })
-    }
-
-    const authorizedWorkspaceId = workflowAuthorization.workflow?.workspaceId
-    if (authorizedWorkspaceId && execution.workspaceId !== authorizedWorkspaceId) {
-      return NextResponse.json({ error: 'Execution not found' }, { status: 404 })
-    }
-
-    const isWorkflowGroupExecution = execution.executionOrigin === 'workflow_group'
-
-    if (execution.status === 'cancelled') {
-      let groupCancellationToPublish: PublishableWorkflowGroupCancellation | null = null
-      let groupCancellationCommitted = false
-      if (isWorkflowGroupExecution) {
-        const preGroupCancellationAbort = cancellationAbortedResponse(abortSignal)
-        if (preGroupCancellationAbort) return preGroupCancellationAbort
-
-        const workflowGroupCancellation = await cancelWorkflowGroupExecution({
-          workspaceId: execution.workspaceId,
-          workflowId,
-          executionId,
-        })
-        if (workflowGroupCancellation.kind === 'conflict') {
-          return NextResponse.json(
-            {
-              error: `Workflow group execution cannot be reconciled while ${workflowGroupCancellation.status}`,
-            },
-            { status: 409 }
-          )
-        }
-        if (workflowGroupCancellation.kind === 'not_workflow_group') {
-          return NextResponse.json(
-            { error: 'Workflow group execution is no longer the active table execution' },
-            { status: 409 }
-          )
-        }
-        if (
-          workflowGroupCancellation.kind === 'cancelled' ||
-          workflowGroupCancellation.kind === 'already_cancelled'
-        ) {
-          groupCancellationToPublish = workflowGroupCancellation
-          groupCancellationCommitted = true
-        }
-      }
-
-      const stopSummary = createExecutionStopSummary()
-      let pausedCancelled = false
-      const pausedCancellationStage = await PauseResumeManager.stagePausedCancellation(
-        executionId,
-        workflowId
-      )
-      const postStageAbort = groupCancellationCommitted
-        ? null
-        : await rollbackPausedCancellationAfterAbort({
-            stage: pausedCancellationStage,
-            workflowId,
-            executionId,
-            abortSignal,
-          })
-      if (postStageAbort) return postStageAbort
-
-      const hasPausedCancellation = isPausedCancellationStage(pausedCancellationStage)
-      const requiresCancellationEvent = hasPausedCancellation || isWorkflowGroupExecution
-      let cancellationEventPublished = !requiresCancellationEvent
-      let activeResumeSignalFailed = false
-      let exactStopSatisfied = true
-      if (pausedCancellationStage.kind === 'active_resume') {
-        exactStopSatisfied = await signalAndRecordActiveResumeStop({
-          workflowId,
-          executionId,
-          executionDeadlineAt: execution.executionDeadlineAt,
-          target: pausedCancellationStage.target,
-          summary: stopSummary,
-        })
-        activeResumeSignalFailed = !exactStopSatisfied
-      } else if (isWorkflowGroupExecution && !hasPausedCancellation) {
-        const retrySignal = await signalExecutionStop({
-          workflowId,
-          signalExecutionId: executionId,
-          executionDeadlineAt: execution.executionDeadlineAt,
-        })
-        mergeExecutionStopSignal(stopSummary, executionId, retrySignal)
-        exactStopSatisfied = retrySignal.accepted
-      }
-
-      if (groupCancellationToPublish && exactStopSatisfied) {
-        await publishWorkflowGroupCancellationEvent(groupCancellationToPublish, executionId)
-      }
-
-      if (requiresCancellationEvent && exactStopSatisfied) {
-        cancellationEventPublished = await ensureCancellationEventPublished(
-          executionId,
-          workflowId,
-          {
-            workspaceId: execution.workspaceId,
-            userId,
-          }
-        )
-      }
-      if (hasPausedCancellation && cancellationEventPublished && exactStopSatisfied) {
-        pausedCancelled = await completePausedCancellationWithRetry(executionId, workflowId, {
-          logMissing: false,
-        })
-      }
-
-      if (exactStopSatisfied) {
-        await releaseExecutionSlot(executionId).catch((error) => {
-          logger.warn('Failed to release reservation while reconciling cancelled execution', {
-            executionId,
-            error: toError(error).message,
-          })
-        })
-      }
-
-      if (pausedCancelled) {
-        await clearStopSignalMarkers(stopSummary)
-      }
-
-      const pausedReconciliationSucceeded =
-        exactStopSatisfied &&
-        (!hasPausedCancellation || (cancellationEventPublished && pausedCancelled))
-      return cancellationOutcome({
-        success: pausedReconciliationSucceeded,
-        executionId,
-        redisAvailable: requiresCancellationEvent ? cancellationEventPublished : true,
-        durablyRecorded: false,
-        locallyAborted: stopSummary.locallyAborted,
-        pausedCancelled,
-        reason: activeResumeSignalFailed
-          ? 'active_resume_signal_failed'
-          : !exactStopSatisfied
-            ? stopSummary.cancellation.reason
-            : hasPausedCancellation && !cancellationEventPublished
-              ? 'paused_event_publish_failed'
-              : hasPausedCancellation && !pausedCancelled
-                ? 'paused_database_cancel_failed'
-                : 'already_cancelled',
-      })
-    }
-
-    if (execution.status !== 'running' && execution.status !== 'pending') {
-      return NextResponse.json(
-        { error: `Execution cannot be cancelled while ${execution.status}` },
-        { status: 409 }
-      )
-    }
-
-    logger.info('Cancel execution requested', { workflowId, executionId, userId })
-
-    const stopSummary = createExecutionStopSummary()
-    let pausedCancelled = false
-    let pausedCancellationStage = await PauseResumeManager.stagePausedCancellation(
-      executionId,
-      workflowId
-    )
-    const postStageAbort = await rollbackPausedCancellationAfterAbort({
-      stage: pausedCancellationStage,
-      workflowId,
-      executionId,
-      abortSignal,
-    })
-    if (postStageAbort) return postStageAbort
-
-    let effectivePausedCancellationPath = isPausedCancellationStage(pausedCancellationStage)
-    let activeResumeTarget =
-      pausedCancellationStage.kind === 'active_resume' ? pausedCancellationStage.target : null
-    let activeResumeEntryId = activeResumeTarget?.resumeEntryId ?? null
-    let activeResumeSignalAccepted = false
-
-    if (activeResumeTarget && !isWorkflowGroupExecution) {
-      activeResumeSignalAccepted = await signalAndRecordActiveResumeStop({
-        workflowId,
-        executionId,
-        executionDeadlineAt: execution.executionDeadlineAt,
-        target: activeResumeTarget,
-        summary: stopSummary,
-      })
-
-      if (!activeResumeSignalAccepted) {
-        const failedResumeEntryId = activeResumeTarget.resumeEntryId
-        await PauseResumeManager.rollbackActiveResumeCancellation(
-          executionId,
-          workflowId,
-          failedResumeEntryId
-        ).catch((error) => {
-          logger.warn('Failed to roll back active resume cancellation intent', {
-            executionId,
-            activeResumeEntryId: failedResumeEntryId,
-            error: toError(error).message,
-          })
-        })
-        await clearStopSignalMarkers(stopSummary)
-        return cancellationOutcome({
-          success: false,
-          executionId,
-          redisAvailable: stopSummary.cancellation.reason !== 'redis_unavailable',
-          durablyRecorded: stopSummary.cancellation.durablyRecorded,
-          locallyAborted: stopSummary.locallyAborted,
-          pausedCancelled: false,
-          reason: 'active_resume_signal_failed',
-        })
-      }
-    } else if (!effectivePausedCancellationPath && !isWorkflowGroupExecution) {
-      const signal = await signalExecutionStop({
-        workflowId,
-        signalExecutionId: executionId,
-        executionDeadlineAt: execution.executionDeadlineAt,
-        queueScope: 'standalone',
-      })
-      mergeExecutionStopSignal(stopSummary, executionId, signal)
-
-      if (!signal.accepted) {
-        pausedCancellationStage = await PauseResumeManager.stagePausedCancellation(
-          executionId,
-          workflowId
-        )
-        const postLateStageAbort = await rollbackPausedCancellationAfterAbort({
-          stage: pausedCancellationStage,
-          workflowId,
-          executionId,
-          abortSignal,
-        })
-        if (postLateStageAbort) {
-          await clearStopSignalMarkers(stopSummary)
-          return postLateStageAbort
-        }
-
-        effectivePausedCancellationPath = isPausedCancellationStage(pausedCancellationStage)
-        activeResumeTarget =
-          pausedCancellationStage.kind === 'active_resume' ? pausedCancellationStage.target : null
-        activeResumeEntryId = activeResumeTarget?.resumeEntryId ?? null
-
-        if (activeResumeTarget) {
-          activeResumeSignalAccepted = await signalAndRecordActiveResumeStop({
-            workflowId,
-            executionId,
-            executionDeadlineAt: execution.executionDeadlineAt,
-            target: activeResumeTarget,
-            summary: stopSummary,
-          })
-          if (!activeResumeSignalAccepted) {
-            const failedResumeEntryId = activeResumeTarget.resumeEntryId
-            await PauseResumeManager.rollbackActiveResumeCancellation(
-              executionId,
-              workflowId,
-              failedResumeEntryId
-            ).catch((error) => {
-              logger.warn('Failed to roll back late active resume cancellation intent', {
-                executionId,
-                activeResumeEntryId: failedResumeEntryId,
-                error: toError(error).message,
-              })
-            })
-            await clearStopSignalMarkers(stopSummary)
-            return cancellationOutcome({
-              success: false,
-              executionId,
-              redisAvailable: stopSummary.cancellation.reason !== 'redis_unavailable',
-              durablyRecorded: stopSummary.cancellation.durablyRecorded,
-              locallyAborted: stopSummary.locallyAborted,
-              pausedCancelled: false,
-              reason: 'active_resume_signal_failed',
-            })
-          }
-        } else if (!effectivePausedCancellationPath) {
-          return cancellationOutcome({
-            success: false,
-            executionId,
-            redisAvailable: stopSummary.cancellation.reason !== 'redis_unavailable',
-            durablyRecorded: stopSummary.cancellation.durablyRecorded,
-            locallyAborted: stopSummary.locallyAborted,
-            pausedCancelled: false,
-            reason: stopSummary.cancellation.reason,
-          })
-        }
-      }
-    }
-
-    let terminalCancellationClaimed = false
-    let competingTerminalStatus: string | null = null
-    let workflowGroupNoLongerActive = false
-    let groupCancellationToPublish: PublishableWorkflowGroupCancellation | null = null
-    try {
-      if (isWorkflowGroupExecution) {
-        const workflowGroupCancellation = await cancelWorkflowGroupExecution({
-          workspaceId: execution.workspaceId,
-          workflowId,
-          executionId,
-        })
-        if (workflowGroupCancellation.kind === 'conflict') {
-          competingTerminalStatus = workflowGroupCancellation.status
-        } else if (workflowGroupCancellation.kind === 'not_workflow_group') {
-          workflowGroupNoLongerActive = true
-        } else {
-          terminalCancellationClaimed = true
-          if (
-            workflowGroupCancellation.kind === 'cancelled' ||
-            workflowGroupCancellation.kind === 'already_cancelled'
-          ) {
-            groupCancellationToPublish = workflowGroupCancellation
-          }
-        }
-      } else {
-        const claim = await claimExecutionLogCancellation({
-          executionId,
-          workflowId,
-          workspaceId: execution.workspaceId,
-        })
-        if (claim.kind === 'cancelled') {
-          terminalCancellationClaimed = true
-        } else {
-          competingTerminalStatus = claim.kind === 'conflict' ? claim.status : 'no_longer_active'
-        }
-      }
-    } catch (dbError) {
-      logger.warn('Failed to finalize cancelled execution directly', {
-        executionId,
-        error: toError(dbError).message,
-      })
-    }
-
-    if (workflowGroupNoLongerActive) {
-      await clearStopSignalMarkers(stopSummary)
-      if (activeResumeEntryId) {
-        await PauseResumeManager.rollbackActiveResumeCancellation(
-          executionId,
-          workflowId,
-          activeResumeEntryId
-        ).catch((error) => {
-          logger.warn('Failed to roll back active resume after group target disappeared', {
-            executionId,
-            activeResumeEntryId,
-            error: toError(error).message,
-          })
-        })
-      } else if (effectivePausedCancellationPath) {
-        await PauseResumeManager.clearPausedCancellationIntent(executionId, workflowId).catch(
-          (error) => {
-            logger.warn('Failed to clear cancellation intent after group target disappeared', {
-              executionId,
-              error: toError(error).message,
-            })
-          }
-        )
-      }
-      return NextResponse.json(
-        { error: 'Workflow group execution is no longer the active table execution' },
-        { status: 409 }
-      )
-    }
-
-    if (competingTerminalStatus) {
-      await clearStopSignalMarkers(stopSummary)
-      if (activeResumeEntryId) {
-        await PauseResumeManager.rollbackActiveResumeCancellation(
-          executionId,
-          workflowId,
-          activeResumeEntryId
-        ).catch((error) => {
-          logger.warn('Failed to roll back active resume after terminal race', {
-            executionId,
-            activeResumeEntryId,
-            error: toError(error).message,
-          })
-        })
-      } else if (effectivePausedCancellationPath) {
-        await PauseResumeManager.clearPausedCancellationIntent(executionId, workflowId).catch(
-          (error) => {
-            logger.warn('Failed to clear cancellation intent after terminal race', {
-              executionId,
-              error: toError(error).message,
-            })
-          }
-        )
-      }
-      return NextResponse.json(
-        {
-          error: isWorkflowGroupExecution
-            ? `Workflow group execution cannot be cancelled while ${competingTerminalStatus}`
-            : `Execution cannot be cancelled while ${competingTerminalStatus}`,
-        },
-        { status: 409 }
-      )
-    }
-
-    if (!terminalCancellationClaimed) {
-      if (effectivePausedCancellationPath && !stopSummary.accepted) {
-        if (activeResumeEntryId) {
-          await PauseResumeManager.rollbackActiveResumeCancellation(
-            executionId,
-            workflowId,
-            activeResumeEntryId
-          )
-        } else {
-          await PauseResumeManager.clearPausedCancellationIntent(executionId, workflowId)
-        }
-      }
-      return cancellationOutcome({
-        success: false,
-        executionId,
-        redisAvailable: stopSummary.cancellation.reason !== 'redis_unavailable',
-        durablyRecorded: stopSummary.cancellation.durablyRecorded,
-        locallyAborted: stopSummary.locallyAborted,
-        pausedCancelled: false,
-        reason: 'cancellation_not_finalized',
-      })
-    }
-
-    let pauseReconciliationFailed = false
-    let activeResumeSignalFailed = false
-    if (isWorkflowGroupExecution) {
-      if (activeResumeTarget) {
-        activeResumeSignalAccepted = await signalAndRecordActiveResumeStop({
-          workflowId,
-          executionId,
-          executionDeadlineAt: execution.executionDeadlineAt,
-          target: activeResumeTarget,
-          summary: stopSummary,
-        })
-        activeResumeSignalFailed = !activeResumeSignalAccepted
-      } else if (!effectivePausedCancellationPath) {
-        const groupSignal = await signalExecutionStop({
-          workflowId,
-          signalExecutionId: executionId,
-          executionDeadlineAt: execution.executionDeadlineAt,
-        })
-        mergeExecutionStopSignal(stopSummary, executionId, groupSignal)
-      }
-    }
-
-    try {
-      const postClaimPausedCancellationStage = await PauseResumeManager.stagePausedCancellation(
-        executionId,
-        workflowId
-      )
-      if (isPausedCancellationStage(postClaimPausedCancellationStage)) {
-        effectivePausedCancellationPath = true
-        if (postClaimPausedCancellationStage.kind === 'active_resume') {
-          const currentActiveResume = postClaimPausedCancellationStage.target
-          const alreadyAttemptedCurrentResume =
-            currentActiveResume.resumeEntryId === activeResumeEntryId &&
-            stopSummary.signalledExecutionIds.has(currentActiveResume.resumeExecutionId)
-          if (!alreadyAttemptedCurrentResume) {
-            activeResumeSignalAccepted = await signalAndRecordActiveResumeStop({
-              workflowId,
-              executionId,
-              executionDeadlineAt: execution.executionDeadlineAt,
-              target: currentActiveResume,
-              summary: stopSummary,
-            })
-          }
-          activeResumeSignalFailed = !activeResumeSignalAccepted
-        } else {
-          activeResumeSignalFailed = false
-        }
-      }
-    } catch (error) {
-      pauseReconciliationFailed = true
-      effectivePausedCancellationPath = true
-      logger.warn('Failed to recheck paused execution after terminal cancellation claim', {
-        executionId,
-        error: toError(error).message,
-      })
-    }
-
-    const executionStopSatisfied = effectivePausedCancellationPath
-      ? !activeResumeSignalFailed
-      : stopSummary.accepted
-    if (groupCancellationToPublish && executionStopSatisfied && !pauseReconciliationFailed) {
-      await publishWorkflowGroupCancellationEvent(groupCancellationToPublish, executionId)
-    }
-    let cancellationEventPublished = false
-    if (executionStopSatisfied && !pauseReconciliationFailed) {
-      cancellationEventPublished = await ensureCancellationEventPublished(executionId, workflowId, {
-        workspaceId: execution.workspaceId,
-        userId,
-      })
-    }
-
-    if (effectivePausedCancellationPath) {
-      if (cancellationEventPublished && !pauseReconciliationFailed && !activeResumeSignalFailed) {
-        pausedCancelled = await completePausedCancellationWithRetry(executionId, workflowId)
-      }
-    } else if (executionStopSatisfied) {
-      await releaseExecutionSlot(executionId).catch((error) => {
-        logger.warn('Failed to release reservation after execution cancellation', {
-          executionId,
-          error: toError(error).message,
-        })
-      })
-    }
-
-    const success = effectivePausedCancellationPath
-      ? pausedCancelled &&
-        cancellationEventPublished &&
-        !pauseReconciliationFailed &&
-        !activeResumeSignalFailed
-      : executionStopSatisfied
-
-    if (effectivePausedCancellationPath && pausedCancelled && cancellationEventPublished) {
-      await clearStopSignalMarkers(stopSummary)
-    }
-
-    if (success) {
-      const workspaceId = execution.workspaceId
-      captureServerEvent(
-        userId,
-        'workflow_execution_cancelled',
-        { workflow_id: workflowId, workspace_id: workspaceId ?? '' },
-        workspaceId ? { groups: { workspace: workspaceId } } : undefined
-      )
-    }
-
-    const durablyRecorded = effectivePausedCancellationPath
-      ? true
-      : stopSummary.cancellation.durablyRecorded
-    const reason = resolveCancellationReason({
-      activeResumeSignalFailed,
-      pauseReconciliationFailed,
-      effectivePausedCancellationPath,
-      cancellationEventPublished,
-      pausedCancelled,
-      stopSummary,
-    })
-
-    return cancellationOutcome({
-      success,
-      executionId,
-      redisAvailable:
-        effectivePausedCancellationPath || pausedCancelled
-          ? cancellationEventPublished
-          : stopSummary.cancellation.reason !== 'redis_unavailable',
-      durablyRecorded,
-      locallyAborted: stopSummary.locallyAborted,
-      pausedCancelled,
-      reason,
-    })
-  } catch (error) {
-    const normalizedError = toError(error)
-    logger.error('Failed to cancel execution', {
-      workflowId,
-      executionId,
-      error: normalizedError.message,
-    })
-    return NextResponse.json(
-      { error: normalizedError.message || 'Failed to cancel execution' },
-      { status: 500 }
-    )
-  }
-}
