+++ apps/sim/lib/execution/cancel-workflow-execution.test.ts
@@ -1,637 +1,1519 @@
 /**
  * @vitest-environment node
  */
+
+import { databaseMock, dbChainMockFns, resetDbChainMock } from '@sim/testing'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 const {
+  mockMarkExecutionCancelled,
+  mockClearExecutionCancellation,
   mockAbortManualExecution,
   mockBeginPausedCancellation,
-  mockBlockQueuedResumes,
-  mockCancelWorkflowGroupExecution,
-  mockCaptureServerEvent,
+  mockStagePausedCancellation,
+  mockBlockQueuedResumesForCancellation,
   mockClearPausedCancellationIntent,
   mockCompletePausedCancellation,
-  mockGetJobQueue,
   mockGetPausedCancellationStatus,
-  mockMarkExecutionCancelled,
-  mockPublishWorkflowGroupCancellationEvent,
+  mockGetActiveResumeCancellationTarget,
+  mockRollbackActiveResumeCancellation,
+  mockFinalizeExecutionStream,
+  mockReadExecutionMetaState,
+  mockWriteEvent,
+  mockWriteTerminalEvent,
+  mockCancelByExecution,
+  mockGetJobQueue,
   mockReleaseExecutionSlot,
-  mockUpdateSet,
-  mockUpdateReturning,
-  mockResolveWorkflowExecutionOwnership,
+  mockCancelWorkflowGroupExecution,
+  mockPublishWorkflowGroupCancellationEvent,
 } = vi.hoisted(() => ({
+  mockMarkExecutionCancelled: vi.fn(),
+  mockClearExecutionCancellation: vi.fn(),
   mockAbortManualExecution: vi.fn(),
   mockBeginPausedCancellation: vi.fn(),
-  mockBlockQueuedResumes: vi.fn(),
-  mockCancelWorkflowGroupExecution: vi.fn(),
-  mockCaptureServerEvent: vi.fn(),
+  mockStagePausedCancellation: vi.fn(),
+  mockBlockQueuedResumesForCancellation: vi.fn(),
   mockClearPausedCancellationIntent: vi.fn(),
   mockCompletePausedCancellation: vi.fn(),
-  mockGetJobQueue: vi.fn(),
   mockGetPausedCancellationStatus: vi.fn(),
-  mockMarkExecutionCancelled: vi.fn(),
-  mockPublishWorkflowGroupCancellationEvent: vi.fn(),
+  mockGetActiveResumeCancellationTarget: vi.fn(),
+  mockRollbackActiveResumeCancellation: vi.fn(),
+  mockFinalizeExecutionStream: vi.fn(),
+  mockReadExecutionMetaState: vi.fn(),
+  mockWriteEvent: vi.fn(),
+  mockWriteTerminalEvent: vi.fn(),
+  mockCancelByExecution: vi.fn(),
+  mockGetJobQueue: vi.fn(),
   mockReleaseExecutionSlot: vi.fn(),
-  mockUpdateSet: vi.fn(),
-  mockUpdateReturning: vi.fn(),
-  mockResolveWorkflowExecutionOwnership: vi.fn(),
-}))
-
-vi.mock('@sim/db', () => ({
-  db: {
-    update: () => ({
-      set: (values: unknown) => {
-        mockUpdateSet(values)
-        return { where: () => ({ returning: () => Promise.resolve(mockUpdateReturning()) }) }
-      },
-    }),
-  },
-}))
-
-vi.mock('@/lib/billing/calculations/usage-reservation', () => ({
-  releaseExecutionSlot: mockReleaseExecutionSlot,
-}))
-
-vi.mock('@/lib/table/workflow-group-cancellation', () => ({
-  cancelWorkflowGroupExecution: mockCancelWorkflowGroupExecution,
-  publishWorkflowGroupCancellationEvent: mockPublishWorkflowGroupCancellationEvent,
-}))
-
-vi.mock('@/lib/execution/cancellation', () => ({
-  markExecutionCancelled: mockMarkExecutionCancelled,
-}))
-
-vi.mock('@/lib/execution/manual-cancellation', () => ({
-  abortManualExecution: mockAbortManualExecution,
-}))
-
-vi.mock('@/lib/execution/event-buffer', () => ({
-  createExecutionEventWriter: () => ({
-    writeTerminal: vi.fn().mockResolvedValue(undefined),
-    close: vi.fn().mockResolvedValue(undefined),
-  }),
-  readExecutionMetaState: vi.fn().mockResolvedValue({ status: 'missing' }),
+  mockCancelWorkflowGroupExecution: vi.fn(),
+  mockPublishWorkflowGroupCancellationEvent: vi.fn(),
 }))
 
 vi.mock('@/lib/core/async-jobs', () => ({
   getJobQueue: mockGetJobQueue,
 }))
 
-vi.mock('@/lib/posthog/server', () => ({
-  captureServerEvent: mockCaptureServerEvent,
+vi.mock('@/lib/billing/calculations/usage-reservation', () => ({
+  releaseExecutionSlot: mockReleaseExecutionSlot,
 }))
 
-vi.mock('@/lib/workflows/executor/execution-job-ids', () => ({
-  WORKFLOW_EXECUTION_JOB_ID_PREFIX: 'workflow-execution:',
+vi.mock('@/lib/execution/cancellation', () => ({
+  markExecutionCancelled: (...args: unknown[]) => mockMarkExecutionCancelled(...args),
+  clearExecutionCancellation: (...args: unknown[]) => mockClearExecutionCancellation(...args),
 }))
 
-vi.mock('@/lib/workflows/executor/execution-queries', () => ({
-  resolveWorkflowExecutionOwnership: mockResolveWorkflowExecutionOwnership,
+vi.mock('@/lib/execution/manual-cancellation', () => ({
+  abortManualExecution: (...args: unknown[]) => mockAbortManualExecution(...args),
 }))
 
 vi.mock('@/lib/workflows/executor/human-in-the-loop-manager', () => ({
   PauseResumeManager: {
-    beginPausedCancellation: mockBeginPausedCancellation,
-    getPausedCancellationStatus: mockGetPausedCancellationStatus,
-    blockQueuedResumesForCancellation: mockBlockQueuedResumes,
-    clearPausedCancellationIntent: mockClearPausedCancellationIntent,
-    completePausedCancellation: mockCompletePausedCancellation,
+    beginPausedCancellation: (...args: unknown[]) => mockBeginPausedCancellation(...args),
+    stagePausedCancellation: (...args: unknown[]) => mockStagePausedCancellation(...args),
+    blockQueuedResumesForCancellation: (...args: unknown[]) =>
+      mockBlockQueuedResumesForCancellation(...args),
+    clearPausedCancellationIntent: (...args: unknown[]) =>
+      mockClearPausedCancellationIntent(...args),
+    completePausedCancellation: (...args: unknown[]) => mockCompletePausedCancellation(...args),
+    getPausedCancellationStatus: (...args: unknown[]) => mockGetPausedCancellationStatus(...args),
+    getActiveResumeCancellationTarget: (...args: unknown[]) =>
+      mockGetActiveResumeCancellationTarget(...args),
+    rollbackActiveResumeCancellation: (...args: unknown[]) =>
+      mockRollbackActiveResumeCancellation(...args),
   },
 }))
 
-import { cancelWorkflowExecution } from '@/lib/execution/cancel-workflow-execution'
+vi.mock('@/lib/table/workflow-group-cancellation', () => ({
+  cancelWorkflowGroupExecution: (...args: unknown[]) => mockCancelWorkflowGroupExecution(...args),
+  publishWorkflowGroupCancellationEvent: (...args: unknown[]) =>
+    mockPublishWorkflowGroupCancellationEvent(...args),
+}))
 
-/**
- * The durable writes a workflow-group transition reports back. The transaction
- * updates the workflow log only, the cell sidecar only, or both, so a single
- * `kind` cannot answer whether this request wrote anything.
- */
-const NO_WRITES = { workflowLogTerminalized: false, sidecarCancelled: false } as const
-const LOG_WRITE = { workflowLogTerminalized: true, sidecarCancelled: false } as const
-const SIDECAR_WRITE = { workflowLogTerminalized: false, sidecarCancelled: true } as const
-const BOTH_WRITES = { workflowLogTerminalized: true, sidecarCancelled: true } as const
+vi.mock('@/lib/execution/event-buffer', () => ({
+  finalizeExecutionStream: (...args: unknown[]) => mockFinalizeExecutionStream(...args),
+  readExecutionMetaState: (...args: unknown[]) => mockReadExecutionMetaState(...args),
+  createExecutionEventWriter: () => ({
+    write: (...args: unknown[]) => mockWriteEvent(...args),
+    writeTerminal: (...args: unknown[]) => mockWriteTerminalEvent(...args),
+    close: vi.fn().mockResolvedValue(undefined),
+  }),
+}))
 
-const INPUT = {
-  executionId: 'execution-1',
-  workflowId: 'workflow-1',
-  userId: 'user-1',
+import { cancelWorkflowExecutionContract } from '@/lib/api/contracts/workflows'
+import { OrchestrationError, statusForOrchestrationError } from '@/lib/core/orchestration/types'
+import {
+  type CancelWorkflowExecutionInput,
+  cancelWorkflowExecution,
+  WorkflowExecutionNotFoundError,
+} from '@/lib/execution/cancel-workflow-execution'
+
+const INPUT: CancelWorkflowExecutionInput = {
+  workflowId: 'wf-1',
+  executionId: 'ex-1',
   workspaceId: 'workspace-1',
+  attributedUserId: 'user-1',
+}
+
+async function cancelAsResponse(
+  overrides: Partial<CancelWorkflowExecutionInput> = {}
+): Promise<Response> {
+  try {
+    const result = await cancelWorkflowExecution({ ...INPUT, ...overrides })
+    const body = cancelWorkflowExecutionContract.response.schema.parse(result)
+    return Response.json(body)
+  } catch (error) {
+    if (error instanceof WorkflowExecutionNotFoundError) {
+      return Response.json({ error: error.message }, { status: 404 })
+    }
+    if (error instanceof OrchestrationError) {
+      return Response.json(
+        { error: error.message },
+        { status: statusForOrchestrationError(error.code) }
+      )
+    }
+    if (error instanceof Error) {
+      return Response.json({ error: error.message }, { status: 500 })
+    }
+    throw error
+  }
+}
+
+const POST = async (..._args: unknown[]) => cancelAsResponse()
+const makeRequest = () => undefined
+const makeParams = () => undefined
+
+const ACTIVE_RESUME_TARGET = {
+  resumeEntryId: 'resume-entry-1',
+  pausedExecutionId: 'paused-1',
+  parentExecutionId: 'ex-1',
+  resumeExecutionId: 'resume-ex-1',
 }
 
 describe('cancelWorkflowExecution', () => {
   beforeEach(() => {
     vi.clearAllMocks()
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: null,
-      priorStatus: 'running',
-    })
-    mockUpdateReturning.mockReturnValue([{ id: 'log-1' }])
-    mockBeginPausedCancellation.mockResolvedValue(false)
-    mockGetPausedCancellationStatus.mockResolvedValue(null)
-    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
-    mockAbortManualExecution.mockReturnValue(false)
-    mockBlockQueuedResumes.mockResolvedValue(undefined)
-    mockClearPausedCancellationIntent.mockResolvedValue(undefined)
-    mockCompletePausedCancellation.mockResolvedValue(true)
-    mockReleaseExecutionSlot.mockResolvedValue(undefined)
-    mockPublishWorkflowGroupCancellationEvent.mockResolvedValue(undefined)
-    mockGetJobQueue.mockResolvedValue({
-      getJob: vi.fn().mockResolvedValue(null),
-      cancelJob: vi.fn(),
-    })
+    resetDbChainMock()
+    dbChainMockFns.limit.mockResolvedValue([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: null,
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    dbChainMockFns.returning.mockResolvedValue([{ status: 'cancelled' }])
+    mockCancelByExecution.mockReset().mockResolvedValue(0)
+    mockGetJobQueue.mockReset().mockResolvedValue({ cancelByExecution: mockCancelByExecution })
+    mockReleaseExecutionSlot.mockReset().mockResolvedValue(undefined)
+    mockCancelWorkflowGroupExecution.mockReset().mockResolvedValue({ kind: 'not_workflow_group' })
+    mockPublishWorkflowGroupCancellationEvent.mockReset().mockResolvedValue(undefined)
+    mockClearExecutionCancellation.mockReset().mockResolvedValue(undefined)
+    mockMarkExecutionCancelled
+      .mockReset()
+      .mockResolvedValue({ durablyRecorded: false, reason: 'redis_unavailable' })
+    mockAbortManualExecution.mockReset().mockReturnValue(false)
+    mockBeginPausedCancellation.mockReset().mockResolvedValue(false)
+    mockStagePausedCancellation.mockReset().mockResolvedValue({ kind: 'not_paused' })
+    mockBlockQueuedResumesForCancellation.mockReset().mockResolvedValue(false)
+    mockClearPausedCancellationIntent.mockReset().mockResolvedValue(undefined)
+    mockCompletePausedCancellation.mockReset().mockResolvedValue(false)
+    mockGetPausedCancellationStatus.mockReset().mockResolvedValue(null)
+    mockGetActiveResumeCancellationTarget.mockReset().mockResolvedValue(null)
+    mockRollbackActiveResumeCancellation.mockReset().mockResolvedValue(true)
+    mockFinalizeExecutionStream.mockReset().mockResolvedValue(true)
+    mockReadExecutionMetaState.mockReset().mockResolvedValue({ status: 'missing' })
+    mockWriteEvent.mockReset().mockResolvedValue({ eventId: 1 })
+    mockWriteTerminalEvent.mockReset().mockResolvedValue({ eventId: 1 })
   })
 
-  /**
-   * The row reads `cancelled` after a successful cancel just as it does after
-   * someone else's, so a status re-read alone would report this run's own work
-   * as `already_cancelled`. Nothing is re-read once the claim moved a row.
-   */
-  it('reports a durable write when an active run is cancelled', async () => {
-    mockResolveWorkflowExecutionOwnership
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: null,
-        priorStatus: 'running',
-      })
-      .mockResolvedValue({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: null,
-        priorStatus: 'cancelled',
-      })
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: true, reason: 'recorded' })
-    expect(mockResolveWorkflowExecutionOwnership).toHaveBeenCalledTimes(1)
-  })
-
-  /**
-   * A cancel against a run that already reached a terminal state changes
-   * nothing: the log claim's `status = 'running'` predicate matches no row and
-   * no terminal metadata moves. Reporting `recorded`/`durablyRecorded: true`
-   * there tells a caller a durable write happened when none did, so the outcome
-   * names the state that was actually observed instead.
-   */
-  it.each([
-    ['cancelled', 'already_cancelled'],
-    ['completed', 'already_completed'],
-    ['failed', 'already_failed'],
-  ])('reports a run already %s as a no-op rather than a durable write', async (status, reason) => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: null,
-      priorStatus: status,
-    })
-    mockUpdateReturning.mockReturnValue([])
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: false, reason })
-  })
-
-  /**
-   * Reclassification only ever applies to an otherwise-clean outcome. A run that
-   * reached any terminal status can still carry paused-HITL state — a
-   * force-failed run keeps whatever pause rows it had — and when reconciling
-   * that genuinely fails, the caller is owed the step that failed rather than a
-   * no-op that also flips `success` to `true`.
-   */
-  it.each([['cancelled'], ['completed'], ['failed']])(
-    'still reports the failing step when a %s run has paused work left over',
-    async (status) => {
-      mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: null,
-        priorStatus: status,
-      })
-      mockBeginPausedCancellation.mockResolvedValue(true)
-      mockCompletePausedCancellation.mockResolvedValue(false)
-
-      const result = await cancelWorkflowExecution(INPUT)
-
-      expect(result).toMatchObject({
-        success: false,
-        durablyRecorded: true,
-        reason: 'paused_database_cancel_failed',
-      })
-    }
-  )
-
-  /**
-   * The status read at entry can be stale: a run that finishes after it and
-   * before the claim leaves a `running` snapshot on a cancel whose claim matched
-   * no row. The claim's own row count is what separates that from a cancel this
-   * request really performed.
-   */
-  it.each([
-    ['completed', 'already_completed'],
-    ['failed', 'already_failed'],
-  ])('reports a run that reached %s after the entry read as a no-op', async (status, reason) => {
-    mockResolveWorkflowExecutionOwnership
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: null,
-        priorStatus: 'running',
-      })
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: null,
-        priorStatus: status,
-      })
-    mockUpdateReturning.mockReturnValue([])
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: false, reason })
-  })
-
-  it('reports an undifferentiated outcome when the claim finds no durable log row', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: null,
-      priorStatus: null,
-    })
-    mockUpdateReturning.mockReturnValue([])
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: true, reason: 'recorded' })
-  })
-
-  /**
-   * The re-read is purely observational — it only refines *which* no-op the
-   * caller is told about. A database that cannot answer it must not take the
-   * cancel down with it: the run has already been cancelled in Redis and its
-   * reservation still has to be released, so the failure degrades to the
-   * undifferentiated outcome rather than propagating.
-   */
-  it('degrades to the undifferentiated outcome when the status re-read fails', async () => {
-    mockResolveWorkflowExecutionOwnership
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: null,
-        priorStatus: 'running',
-      })
-      .mockRejectedValueOnce(new Error('connection terminated'))
-    mockUpdateReturning.mockReturnValue([])
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: true, reason: 'recorded' })
-    expect(mockResolveWorkflowExecutionOwnership).toHaveBeenCalledTimes(2)
-    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('execution-1')
-  })
-
-  it('releases the plan concurrency reservation after a successful cancellation', async () => {
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result.success).toBe(true)
-    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('execution-1')
-  })
-
-  it('keeps the reservation held when nothing could be cancelled', async () => {
+  it('returns success when cancellation was durably recorded', async () => {
     mockMarkExecutionCancelled.mockResolvedValue({
-      durablyRecorded: false,
-      reason: 'redis_unavailable',
+      durablyRecorded: true,
+      reason: 'recorded',
     })
 
-    const result = await cancelWorkflowExecution(INPUT)
+    const response = await POST(makeRequest(), makeParams())
 
-    expect(result.success).toBe(false)
-    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      success: true,
+      executionId: 'ex-1',
+      redisAvailable: true,
+      durablyRecorded: true,
+      locallyAborted: false,
+      pausedCancelled: false,
+      reason: 'recorded',
+    })
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      {
+        workflowId: 'wf-1',
+        executionId: 'ex-1',
+      },
+      'standalone'
+    )
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1', {
+      executionDeadlineAt: null,
+    })
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
   })
 
-  it('cancels and publishes the table cell sidecar of a workflow-group execution', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'running',
-    })
-    const cancelled = {
-      kind: 'cancelled' as const,
+  it('atomically claims one workflow-group attempt before signalling it', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
       tableId: 'table-1',
       rowId: 'row-1',
       groupId: 'group-1',
-      writes: BOTH_WRITES,
-    }
-    mockCancelWorkflowGroupExecution.mockResolvedValue(cancelled)
+    })
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: true,
+      reason: 'recorded',
+    })
 
-    const result = await cancelWorkflowExecution(INPUT)
+    const response = await POST(makeRequest(), makeParams())
 
-    expect(result.success).toBe(true)
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      executionId: 'ex-1',
+      redisAvailable: true,
+      durablyRecorded: true,
+      reason: 'recorded',
+    })
     expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledWith({
       workspaceId: 'workspace-1',
-      workflowId: 'workflow-1',
-      executionId: 'execution-1',
+      workflowId: 'wf-1',
+      executionId: 'ex-1',
     })
-    expect(mockPublishWorkflowGroupCancellationEvent).toHaveBeenCalledWith(cancelled, 'execution-1')
-    expect(mockUpdateSet).not.toHaveBeenCalled()
-    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('execution-1')
+    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
+      mockMarkExecutionCancelled.mock.invocationCallOrder[0]
+    )
+    expect(mockMarkExecutionCancelled.mock.invocationCallOrder[0]).toBeLessThan(
+      mockPublishWorkflowGroupCancellationEvent.mock.invocationCallOrder[0]
+    )
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+    expect(mockStagePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
+    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
   })
 
-  /**
-   * A workflow-group log that is already `cancelled` can still own a cell
-   * sidecar left in `error`, and reconciling it to `cancelled` is a durable
-   * write this request performed. The terminal entry snapshot cannot see that
-   * work, so it must not reinterpret the outcome as a no-op — the API would
-   * otherwise tell the caller nothing changed and drop the cancellation event.
-   */
-  it('reports a durable write when a cancelled group run still had its sidecar reconciled', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'cancelled',
-    })
-    const cancelled = {
-      kind: 'cancelled' as const,
+  it('keeps a claimed group cancellation retryable when signalling fails', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
       tableId: 'table-1',
       rowId: 'row-1',
       groupId: 'group-1',
-      writes: SIDECAR_WRITE,
-    }
-    mockCancelWorkflowGroupExecution.mockResolvedValue(cancelled)
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: true, reason: 'recorded' })
-    expect(mockPublishWorkflowGroupCancellationEvent).toHaveBeenCalledWith(cancelled, 'execution-1')
-  })
-
-  /**
-   * The group path terminalizes the workflow log itself when the cell sidecar
-   * is already gone, so that outcome is a durable write too.
-   */
-  it('reports a durable write when the group path cancels a run whose sidecar is gone', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'running',
-    })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'cancelled_without_sidecar',
-      writes: LOG_WRITE,
-    })
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: true, reason: 'recorded' })
-    expect(mockUpdateSet).not.toHaveBeenCalled()
-  })
-
-  /**
-   * The mirror case: a group run that was already terminal and whose sidecar was
-   * already `cancelled` leaves both records untouched, so it still reports the
-   * state it observed rather than a durable write.
-   */
-  it('reports a group run that changed nothing as a no-op', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'cancelled',
-    })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'already_cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-      writes: NO_WRITES,
-    })
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({
-      success: true,
-      durablyRecorded: false,
-      reason: 'already_cancelled',
-    })
-  })
-
-  /**
-   * The same `already_cancelled` kind covers a transition that left the sidecar
-   * alone but still terminalized an active workflow log. That log write is
-   * durable, so the outcome must stay `recorded` and must not re-read a state
-   * this request itself wrote.
-   */
-  it('reports a durable write when a group run only repaired its workflow log', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'running',
-    })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'already_cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-      writes: LOG_WRITE,
-    })
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({ success: true, durablyRecorded: true, reason: 'recorded' })
-    expect(mockResolveWorkflowExecutionOwnership).toHaveBeenCalledOnce()
-  })
-
-  /**
-   * The lost race the sidecar-bearing kind used to hide: a concurrent cancel
-   * terminalized both records between the entry snapshot and this transaction,
-   * which then found the sidecar already `cancelled` and the log already
-   * `cancelled` and wrote nothing. A non-terminal entry snapshot cannot catch
-   * that, so the transition's own report of having written nothing is what
-   * forces the re-read — otherwise the request would claim a durable write and
-   * fire the v2 cancel analytics gate on a no-op.
-   */
-  it('reports a group run that lost the race with its sidecar already cancelled as a no-op', async () => {
-    mockResolveWorkflowExecutionOwnership
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: 'workspace-1',
-        priorStatus: 'running',
-      })
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: 'workspace-1',
-        priorStatus: 'cancelled',
-      })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'already_cancelled',
-      tableId: 'table-1',
-      rowId: 'row-1',
-      groupId: 'group-1',
-      writes: NO_WRITES,
-    })
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({
-      success: true,
-      durablyRecorded: false,
-      reason: 'already_cancelled',
-    })
-    expect(mockResolveWorkflowExecutionOwnership).toHaveBeenCalledTimes(2)
-  })
-
-  /**
-   * The lost-race re-read applies to the group path as well: an entry snapshot
-   * can still read `running` when the sidecar-less transition finds the log
-   * already `cancelled` and writes nothing.
-   */
-  it('reports a group run that lost the race to another cancel as a no-op', async () => {
-    mockResolveWorkflowExecutionOwnership
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: 'workspace-1',
-        priorStatus: 'running',
-      })
-      .mockResolvedValueOnce({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: 'workspace-1',
-        priorStatus: 'cancelled',
-      })
-    mockCancelWorkflowGroupExecution.mockResolvedValue({
-      kind: 'already_cancelled_without_sidecar',
-      writes: NO_WRITES,
-    })
-
-    const result = await cancelWorkflowExecution(INPUT)
-
-    expect(result).toMatchObject({
-      success: true,
-      durablyRecorded: false,
-      reason: 'already_cancelled',
-    })
-    expect(mockResolveWorkflowExecutionOwnership).toHaveBeenCalledTimes(2)
-  })
-
-  it.each([
-    [
-      { kind: 'conflict' as const, status: 'completed', writes: NO_WRITES },
-      'cannot be cancelled while completed',
-    ],
-    [
-      { kind: 'not_workflow_group' as const, writes: NO_WRITES },
-      'no longer the active table execution',
-    ],
-  ])(
-    'releases the reservation before reporting a refused workflow-group cell claim as a conflict',
-    async (outcome, message) => {
-      mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: 'workspace-1',
-      })
-      mockCancelWorkflowGroupExecution.mockResolvedValue(outcome)
-
-      await expect(cancelWorkflowExecution(INPUT)).rejects.toMatchObject({
-        code: 'conflict',
-        message: expect.stringContaining(message),
-      })
-      expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
-      expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('execution-1')
-    }
-  )
-
-  it.each([
-    [{ kind: 'conflict' as const, status: 'completed', writes: NO_WRITES }],
-    [{ kind: 'not_workflow_group' as const, writes: NO_WRITES }],
-  ])(
-    'keeps the reservation held when a refused claim follows a paused cancellation',
-    async (outcome) => {
-      mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: 'workspace-1',
-      })
-      mockBeginPausedCancellation.mockResolvedValue(true)
-      mockCancelWorkflowGroupExecution.mockResolvedValue(outcome)
-
-      await expect(cancelWorkflowExecution(INPUT)).rejects.toMatchObject({ code: 'conflict' })
-      expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-    }
-  )
-
-  it.each([
-    [{ kind: 'conflict' as const, status: 'completed', writes: NO_WRITES }],
-    [{ kind: 'not_workflow_group' as const, writes: NO_WRITES }],
-  ])(
-    'keeps the reservation held when a refused claim follows a failed cancellation',
-    async (outcome) => {
-      mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-        belongsToWorkflow: true,
-        workflowGroupWorkspaceId: 'workspace-1',
-      })
-      mockMarkExecutionCancelled.mockResolvedValue({
-        durablyRecorded: false,
-        reason: 'redis_unavailable',
-      })
-      mockCancelWorkflowGroupExecution.mockResolvedValue(outcome)
-
-      await expect(cancelWorkflowExecution(INPUT)).rejects.toMatchObject({ code: 'conflict' })
-      expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-    }
-  )
-
-  it('releases the reservation and rethrows when the workflow-group cancel fails unexpectedly', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'running',
-    })
-    const failure = new Error('Workflow-group cancellation lost its locked workflow-log claim')
-    mockCancelWorkflowGroupExecution.mockRejectedValue(failure)
-
-    await expect(cancelWorkflowExecution(INPUT)).rejects.toBe(failure)
-    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('execution-1')
-    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
-    expect(mockUpdateSet).not.toHaveBeenCalled()
-  })
-
-  it('keeps the reservation held when an unexpected workflow-group failure follows a paused cancellation', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'running',
-    })
-    mockBeginPausedCancellation.mockResolvedValue(true)
-    mockCancelWorkflowGroupExecution.mockRejectedValue(new Error('serialization conflict'))
-
-    await expect(cancelWorkflowExecution(INPUT)).rejects.toThrow('serialization conflict')
-    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
-  })
-
-  it('keeps the reservation held when an unexpected workflow-group failure follows a failed cancellation', async () => {
-    mockResolveWorkflowExecutionOwnership.mockResolvedValue({
-      belongsToWorkflow: true,
-      workflowGroupWorkspaceId: 'workspace-1',
-      priorStatus: 'running',
     })
     mockMarkExecutionCancelled.mockResolvedValue({
       durablyRecorded: false,
       reason: 'redis_unavailable',
     })
-    mockCancelWorkflowGroupExecution.mockRejectedValue(new Error('serialization conflict'))
 
-    await expect(cancelWorkflowExecution(INPUT)).rejects.toThrow('serialization conflict')
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: false,
+      durablyRecorded: false,
+      reason: 'redis_unavailable',
+    })
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledOnce()
+    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
+    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
+    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+  })
+
+  it('accepts an exact in-process group abort without cancelling its carrier', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: false,
+      reason: 'redis_unavailable',
+    })
+    mockAbortManualExecution.mockReturnValue(true)
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      durablyRecorded: false,
+      locallyAborted: true,
+      reason: 'redis_unavailable',
+    })
+    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+  })
+
+  it('does not cancel a shared carrier when an idle workflow-group pause appears', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockStagePausedCancellation
+      .mockResolvedValueOnce({ kind: 'not_paused' })
+      .mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      pausedCancelled: true,
+      reason: 'recorded',
+    })
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+    expect(mockStagePausedCancellation.mock.invocationCallOrder[1]).toBeLessThan(
+      mockWriteTerminalEvent.mock.invocationCallOrder[0]
+    )
+    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
+  })
+
+  it('leaves a late workflow-group pause retryable when event publication fails', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockStagePausedCancellation
+      .mockResolvedValueOnce({ kind: 'not_paused' })
+      .mockResolvedValue({ kind: 'idle' })
+    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: false,
+      pausedCancelled: false,
+      reason: 'paused_event_publish_failed',
+    })
+    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
+    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
+      mockWriteTerminalEvent.mock.invocationCallOrder[0]
+    )
+    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
+    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
+  })
+
+  it('cancels an active workflow-group resume without cancelling its shared carrier', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockStagePausedCancellation.mockResolvedValue({
+      kind: 'active_resume',
+      target: ACTIVE_RESUME_TARGET,
+    })
+    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: false,
+      reason: 'redis_unavailable',
+    })
+    mockCancelByExecution.mockResolvedValue(1)
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      pausedCancelled: true,
+      reason: 'recorded',
+    })
+    expect(mockCancelByExecution).toHaveBeenCalledOnce()
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      { workflowId: 'wf-1', executionId: 'ex-1' },
+      'resume'
+    )
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('resume-ex-1', {
+      executionDeadlineAt: null,
+    })
+    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
+    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
+      mockMarkExecutionCancelled.mock.invocationCallOrder[0]
+    )
+    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
+    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
+      mockCompletePausedCancellation.mock.invocationCallOrder[0]
+    )
+  })
+
+  it('keeps a claimed active group resume retryable when no stop backend accepts it', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockStagePausedCancellation.mockResolvedValue({
+      kind: 'active_resume',
+      target: ACTIVE_RESUME_TARGET,
+    })
+    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: false,
+      durablyRecorded: true,
+      locallyAborted: false,
+      pausedCancelled: false,
+      reason: 'active_resume_signal_failed',
+    })
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('resume-ex-1', {
+      executionDeadlineAt: null,
+    })
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      { workflowId: 'wf-1', executionId: 'ex-1' },
+      'resume'
+    )
+    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
+      mockMarkExecutionCancelled.mock.invocationCallOrder[0]
+    )
+    expect(mockRollbackActiveResumeCancellation).not.toHaveBeenCalled()
+    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
+  })
+
+  it('rechecks for a pause after the group terminal claim waits on persistence', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockStagePausedCancellation
+      .mockResolvedValueOnce({ kind: 'not_paused' })
+      .mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      pausedCancelled: true,
+      reason: 'recorded',
+    })
+    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
+      mockStagePausedCancellation.mock.invocationCallOrder[1]
+    )
+    expect(mockStagePausedCancellation.mock.invocationCallOrder[1]).toBeLessThan(
+      mockWriteTerminalEvent.mock.invocationCallOrder[0]
+    )
+    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
+      mockCompletePausedCancellation.mock.invocationCallOrder[0]
+    )
+    expect(mockWriteTerminalEvent).toHaveBeenCalledOnce()
+  })
+
+  it('rechecks for a regular pause after the terminal claim waits on persistence', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockStagePausedCancellation
+      .mockResolvedValueOnce({ kind: 'not_paused' })
+      .mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      pausedCancelled: true,
+      reason: 'recorded',
+    })
+    expect(databaseMock.db.update.mock.invocationCallOrder[0]).toBeLessThan(
+      mockStagePausedCancellation.mock.invocationCallOrder[1]
+    )
+    expect(mockStagePausedCancellation.mock.invocationCallOrder[1]).toBeLessThan(
+      mockWriteTerminalEvent.mock.invocationCallOrder[0]
+    )
+    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
+      mockCompletePausedCancellation.mock.invocationCallOrder[0]
+    )
+  })
+
+  it('uses generic cancellation for a regular Table-trigger-block execution', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: null,
+        status: 'running',
+        trigger: 'table',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: true,
+      reason: 'recorded',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    expect(mockCancelWorkflowGroupExecution).not.toHaveBeenCalled()
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1', {
+      executionDeadlineAt: null,
+    })
+    expect(mockCancelByExecution).toHaveBeenCalledOnce()
+  })
+
+  it('returns 409 without generic cancellation when the exact group attempt is stale', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'conflict',
+      status: 'no_longer_active',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(409)
+    await expect(response.json()).resolves.toEqual({
+      error: 'Workflow group execution cannot be cancelled while no_longer_active',
+    })
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+    expect(mockStagePausedCancellation).toHaveBeenCalledOnce()
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
+    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
+  })
+
+  it('terminalizes a durable workflow-group log after its table sidecar was deleted', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockCancelWorkflowGroupExecution.mockResolvedValue({ kind: 'cancelled_without_sidecar' })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      executionId: 'ex-1',
+      reason: 'recorded',
+    })
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledOnce()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
+    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
+  })
+
+  it('returns unsuccessful response when Redis is unavailable', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: false,
+      reason: 'redis_unavailable',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      success: false,
+      executionId: 'ex-1',
+      redisAvailable: false,
+      durablyRecorded: false,
+      locallyAborted: false,
+      pausedCancelled: false,
+      reason: 'redis_unavailable',
+    })
+  })
+
+  it('returns unsuccessful response when Redis persistence fails', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: false,
+      reason: 'redis_write_failed',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      success: false,
+      executionId: 'ex-1',
+      redisAvailable: true,
+      durablyRecorded: false,
+      locallyAborted: false,
+      pausedCancelled: false,
+      reason: 'redis_write_failed',
+    })
+  })
+
+  it('returns success when local fallback aborts execution without Redis durability', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: false,
+      reason: 'redis_unavailable',
+    })
+    mockAbortManualExecution.mockReturnValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      success: true,
+      executionId: 'ex-1',
+      redisAvailable: false,
+      durablyRecorded: false,
+      locallyAborted: true,
+      pausedCancelled: false,
+      reason: 'redis_unavailable',
+    })
+  })
+
+  it('returns success when the queue backend cancels the active job', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: false,
+      reason: 'redis_unavailable',
+    })
+    mockCancelByExecution.mockResolvedValue(1)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      executionId: 'ex-1',
+      durablyRecorded: false,
+      locallyAborted: false,
+      reason: 'queue_cancelled',
+    })
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
+  })
+
+  it('cancels a queued execution before its workflow log exists', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([])
+    mockCancelByExecution.mockResolvedValueOnce(1)
+    mockMarkExecutionCancelled.mockResolvedValueOnce({
+      durablyRecorded: true,
+      reason: 'recorded',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      success: true,
+      executionId: 'ex-1',
+      redisAvailable: true,
+      durablyRecorded: true,
+      locallyAborted: false,
+      pausedCancelled: false,
+      reason: 'queue_cancelled',
+    })
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      {
+        workflowId: 'wf-1',
+        executionId: 'ex-1',
+      },
+      'standalone'
+    )
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1')
+    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('ex-1')
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
+  })
+
+  it('does not use an unscoped local abort before its workflow log exists', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([])
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(404)
+    await expect(response.json()).resolves.toEqual({ error: 'Execution not found' })
+    expect(mockAbortManualExecution).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      {
+        workflowId: 'wf-1',
+        executionId: 'ex-1',
+      },
+      'standalone'
+    )
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
+  })
+
+  it('returns success when a paused HITL execution is cancelled directly in the database', async () => {
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      success: true,
+      executionId: 'ex-1',
+      redisAvailable: true,
+      durablyRecorded: true,
+      locallyAborted: false,
+      pausedCancelled: true,
+      reason: 'recorded',
+    })
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockWriteTerminalEvent).toHaveBeenCalledWith(
+      expect.objectContaining({
+        type: 'execution:cancelled',
+        executionId: 'ex-1',
+        workflowId: 'wf-1',
+      }),
+      'cancelled'
+    )
+    expect(mockFinalizeExecutionStream).not.toHaveBeenCalled()
+  })
+
+  it('claims the paused workflow-group sidecar before publishing and finalizing', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      durablyRecorded: true,
+      pausedCancelled: true,
+      reason: 'recorded',
+    })
+    expect(mockStagePausedCancellation.mock.invocationCallOrder[0]).toBeLessThan(
+      mockWriteTerminalEvent.mock.invocationCallOrder[0]
+    )
+    expect(mockCancelWorkflowGroupExecution.mock.invocationCallOrder[0]).toBeLessThan(
+      mockWriteTerminalEvent.mock.invocationCallOrder[0]
+    )
+    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
+      mockCompletePausedCancellation.mock.invocationCallOrder[0]
+    )
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+    expect(mockWriteTerminalEvent).toHaveBeenCalledOnce()
+    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
+  })
+
+  it('keeps a paused workflow-group cancellation reserved when event publication fails', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'running',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: false,
+      pausedCancelled: false,
+      reason: 'paused_event_publish_failed',
+    })
+    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
+    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledOnce()
+    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+  })
+
+  it('publishes paused cancellation event even when Redis cancellation is recorded', async () => {
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      executionId: 'ex-1',
+      durablyRecorded: true,
+      pausedCancelled: true,
+    })
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockWriteTerminalEvent).toHaveBeenCalledWith(
+      expect.objectContaining({
+        type: 'execution:cancelled',
+        executionId: 'ex-1',
+        workflowId: 'wf-1',
+      }),
+      'cancelled'
+    )
+    expect(mockFinalizeExecutionStream).not.toHaveBeenCalled()
+  })
+
+  it('does not confirm paused cancellation when terminal event publication fails', async () => {
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      success: false,
+      executionId: 'ex-1',
+      redisAvailable: false,
+      durablyRecorded: true,
+      locallyAborted: false,
+      pausedCancelled: false,
+      reason: 'paused_event_publish_failed',
+    })
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
+    expect(mockWriteTerminalEvent).toHaveBeenCalledWith(
+      expect.objectContaining({
+        type: 'execution:cancelled',
+        executionId: 'ex-1',
+        workflowId: 'wf-1',
+      }),
+      'cancelled'
+    )
+    expect(mockFinalizeExecutionStream).not.toHaveBeenCalled()
+  })
+
+  it('finishes reconciliation when the pause row is already cancelled', async () => {
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      pausedCancelled: true,
+      reason: 'recorded',
+    })
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
+      mockCompletePausedCancellation.mock.invocationCallOrder[0]
+    )
+  })
+
+  it('stops before lookup when cancellation is already aborted', async () => {
+    const controller = new AbortController()
+    controller.abort()
+
+    const response = await cancelAsResponse({
+      abortSignal: controller.signal,
+    })
+
+    expect(response.status).toBe(409)
+    await expect(response.json()).resolves.toEqual({
+      error: 'Request aborted before workflow run cancellation could be applied.',
+    })
+    expect(databaseMock.db.select).not.toHaveBeenCalled()
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+  })
+
+  it('stops before mutation when cancellation is aborted during execution lookup', async () => {
+    const controller = new AbortController()
+    dbChainMockFns.limit.mockImplementationOnce(async () => {
+      controller.abort()
+      return [
+        {
+          executionDeadlineAt: null,
+          executionOrigin: null,
+          status: 'running',
+          workspaceId: 'workspace-1',
+        },
+      ]
+    })
+
+    const response = await cancelAsResponse({
+      abortSignal: controller.signal,
+    })
+
+    expect(response.status).toBe(409)
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockAbortManualExecution).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+  })
+
+  it('rolls back pause staging when cancellation is aborted during staging', async () => {
+    const controller = new AbortController()
+    mockStagePausedCancellation.mockImplementationOnce(async () => {
+      controller.abort()
+      return { kind: 'idle' }
+    })
+
+    const response = await cancelAsResponse({
+      abortSignal: controller.signal,
+    })
+
+    expect(response.status).toBe(409)
+    expect(mockClearPausedCancellationIntent).toHaveBeenCalledWith('ex-1', 'wf-1')
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockAbortManualExecution).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+  })
+
+  it('rolls back an active resume staged while cancellation is aborted', async () => {
+    const controller = new AbortController()
+    mockStagePausedCancellation.mockImplementationOnce(async () => {
+      controller.abort()
+      return { kind: 'active_resume', target: ACTIVE_RESUME_TARGET }
+    })
+
+    const response = await cancelAsResponse({
+      abortSignal: controller.signal,
+    })
+
+    expect(response.status).toBe(409)
+    expect(mockRollbackActiveResumeCancellation).toHaveBeenCalledWith(
+      'ex-1',
+      'wf-1',
+      'resume-entry-1'
+    )
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+  })
+
+  it('rolls back pause staging for an already-cancelled execution when cancellation is aborted', async () => {
+    const controller = new AbortController()
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: null,
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockStagePausedCancellation.mockImplementationOnce(async () => {
+      controller.abort()
+      return { kind: 'idle' }
+    })
+
+    const response = await cancelAsResponse({
+      abortSignal: controller.signal,
+    })
+
+    expect(response.status).toBe(409)
+    expect(mockClearPausedCancellationIntent).toHaveBeenCalledWith('ex-1', 'wf-1')
+    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
+    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
     expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
   })
 
-  it('leaves a standalone execution untouched by the workflow-group path', async () => {
-    await cancelWorkflowExecution(INPUT)
+  it('finishes cancellation when an aborted active-resume stage cannot be rolled back', async () => {
+    const controller = new AbortController()
+    mockStagePausedCancellation.mockImplementationOnce(async () => {
+      controller.abort()
+      return { kind: 'active_resume', target: ACTIVE_RESUME_TARGET }
+    })
+    mockRollbackActiveResumeCancellation.mockResolvedValueOnce(false)
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
 
-    expect(mockCancelWorkflowGroupExecution).not.toHaveBeenCalled()
-    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
-    expect(mockUpdateSet).toHaveBeenCalledWith(expect.objectContaining({ status: 'cancelled' }))
+    const response = await cancelAsResponse({
+      abortSignal: controller.signal,
+    })
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({ success: true, pausedCancelled: true })
+    expect(mockRollbackActiveResumeCancellation).toHaveBeenCalledWith(
+      'ex-1',
+      'wf-1',
+      'resume-entry-1'
+    )
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('resume-ex-1', {
+      executionDeadlineAt: null,
+    })
   })
 
-  /**
-   * A cancelled run is terminal, so it owes the same two fields every other
-   * terminal write records. Without the duration it is invisible to the
-   * `minDurationMs`/`maxDurationMs` filters on `GET /api/v2/logs`.
-   */
-  it('records how long the cancelled run had been going, not just when it stopped', async () => {
-    await cancelWorkflowExecution(INPUT)
+  it('returns 404 when the execution does not belong to the workflow', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([])
 
-    const [values] = mockUpdateSet.mock.calls.at(-1) as [Record<string, unknown>]
-    expect(values.endedAt).toBeInstanceOf(Date)
-    expect(values.totalDurationMs).toBeDefined()
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(404)
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      {
+        workflowId: 'wf-1',
+        executionId: 'ex-1',
+      },
+      'standalone'
+    )
+  })
+
+  it('treats an already-cancelled execution as an idempotent success', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      { executionDeadlineAt: null, status: 'cancelled', workspaceId: 'workspace-1' },
+    ])
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      durablyRecorded: false,
+      reason: 'already_cancelled',
+    })
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('ex-1')
+  })
+
+  it('reconciles the exact sidecar when a workflow-group log is already cancelled', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'already_cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      reason: 'already_cancelled',
+    })
+    expect(mockCancelWorkflowGroupExecution).toHaveBeenCalledWith({
+      workspaceId: 'workspace-1',
+      workflowId: 'wf-1',
+      executionId: 'ex-1',
+    })
+    expect(mockMarkExecutionCancelled).toHaveBeenCalledWith('ex-1', {
+      executionDeadlineAt: null,
+    })
+    expect(mockPublishWorkflowGroupCancellationEvent).toHaveBeenCalledOnce()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+  })
+
+  it('finishes workflow-group reconciliation when abort arrives during its durable commit', async () => {
+    const controller = new AbortController()
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockImplementationOnce(async () => {
+      controller.abort()
+      return {
+        kind: 'already_cancelled',
+        tableId: 'table-1',
+        rowId: 'row-1',
+        groupId: 'group-1',
+      }
+    })
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await cancelAsResponse({
+      abortSignal: controller.signal,
+    })
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({ success: true, pausedCancelled: true })
+    expect(mockClearPausedCancellationIntent).not.toHaveBeenCalled()
+    expect(mockPublishWorkflowGroupCancellationEvent).toHaveBeenCalledOnce()
+    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
+  })
+
+  it('does not finalize an already-cancelled group retry until exact stop is accepted', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'already_cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: false,
+      redisAvailable: false,
+      reason: 'redis_unavailable',
+    })
+    expect(mockPublishWorkflowGroupCancellationEvent).not.toHaveBeenCalled()
+    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
+    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
+  })
+
+  it('repairs a stranded active resume when the group log is already cancelled', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockStagePausedCancellation.mockResolvedValue({
+      kind: 'active_resume',
+      target: ACTIVE_RESUME_TARGET,
+    })
+    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
+    mockAbortManualExecution.mockReturnValue(true)
+    mockCancelByExecution.mockResolvedValue(1)
+    mockCompletePausedCancellation.mockResolvedValue(true)
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'already_cancelled',
+      tableId: 'table-1',
+      rowId: 'row-1',
+      groupId: 'group-1',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      locallyAborted: true,
+      pausedCancelled: true,
+      reason: 'already_cancelled',
+    })
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      { workflowId: 'wf-1', executionId: 'ex-1' },
+      'resume'
+    )
+    expect(mockAbortManualExecution).toHaveBeenCalledWith('resume-ex-1')
+    expect(mockCompletePausedCancellation).toHaveBeenCalledWith('ex-1', 'wf-1')
+    expect(mockWriteTerminalEvent).toHaveBeenCalledOnce()
+    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
+      mockCompletePausedCancellation.mock.invocationCallOrder[0]
+    )
+  })
+
+  it('repairs a stranded active resume when a regular workflow log is already cancelled', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: null,
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockStagePausedCancellation.mockResolvedValue({
+      kind: 'active_resume',
+      target: ACTIVE_RESUME_TARGET,
+    })
+    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
+    mockAbortManualExecution.mockReturnValue(true)
+    mockCancelByExecution.mockResolvedValue(1)
+    mockCompletePausedCancellation.mockResolvedValue(true)
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: true,
+      locallyAborted: true,
+      pausedCancelled: true,
+      reason: 'already_cancelled',
+    })
+    expect(mockCancelByExecution).toHaveBeenCalledWith(
+      { workflowId: 'wf-1', executionId: 'ex-1' },
+      'resume'
+    )
+    expect(mockWriteTerminalEvent.mock.invocationCallOrder[0]).toBeLessThan(
+      mockCompletePausedCancellation.mock.invocationCallOrder[0]
+    )
+    expect(mockCancelWorkflowGroupExecution).not.toHaveBeenCalled()
+  })
+
+  it('keeps an already-cancelled active resume retryable when publication fails', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: null,
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockStagePausedCancellation.mockResolvedValue({
+      kind: 'active_resume',
+      target: ACTIVE_RESUME_TARGET,
+    })
+    mockGetActiveResumeCancellationTarget.mockResolvedValue(ACTIVE_RESUME_TARGET)
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    mockWriteTerminalEvent.mockRejectedValue(new Error('Redis unavailable'))
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({
+      success: false,
+      pausedCancelled: false,
+      reason: 'paused_event_publish_failed',
+    })
+    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
+  })
+
+  it('fails closed when an already-cancelled group sidecar cannot be reconciled', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockResolvedValue({
+      kind: 'conflict',
+      status: 'completed',
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(409)
+    await expect(response.json()).resolves.toEqual({
+      error: 'Workflow group execution cannot be reconciled while completed',
+    })
+  })
+
+  it('reports a reconciliation failure for an already-cancelled group sidecar', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      {
+        executionDeadlineAt: null,
+        executionOrigin: 'workflow_group',
+        status: 'cancelled',
+        workspaceId: 'workspace-1',
+      },
+    ])
+    mockCancelWorkflowGroupExecution.mockRejectedValue(new Error('database unavailable'))
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(500)
+    await expect(response.json()).resolves.toEqual({ error: 'database unavailable' })
+  })
+
+  it('returns 409 when the execution is already terminal', async () => {
+    dbChainMockFns.limit.mockResolvedValueOnce([
+      { status: 'completed', workspaceId: 'workspace-1' },
+    ])
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(409)
+    expect(mockMarkExecutionCancelled).not.toHaveBeenCalled()
+    expect(mockCancelByExecution).not.toHaveBeenCalled()
+  })
+
+  it('returns 409 when completion wins the terminal database race', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    const returning = vi.fn().mockResolvedValue([])
+    const where = vi.fn(() => ({ returning }))
+    databaseMock.db.update.mockReturnValueOnce({ set: vi.fn(() => ({ where })) })
+    dbChainMockFns.limit
+      .mockResolvedValueOnce([
+        { executionDeadlineAt: null, status: 'running', workspaceId: 'workspace-1' },
+      ])
+      .mockResolvedValueOnce([{ status: 'completed' }])
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(409)
+    await expect(response.json()).resolves.toEqual({
+      error: 'Execution cannot be cancelled while completed',
+    })
+    expect(returning).toHaveBeenCalledOnce()
+    expect(mockClearExecutionCancellation).toHaveBeenCalledWith('ex-1')
+    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
+    expect(mockReleaseExecutionSlot).not.toHaveBeenCalled()
+  })
+
+  it('rolls back paused cancellation intent when resume completion wins the log claim', async () => {
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+    const returning = vi.fn().mockResolvedValue([])
+    const where = vi.fn(() => ({ returning }))
+    databaseMock.db.update.mockReturnValueOnce({ set: vi.fn(() => ({ where })) })
+    dbChainMockFns.limit
+      .mockResolvedValueOnce([
+        { executionDeadlineAt: null, status: 'running', workspaceId: 'workspace-1' },
+      ])
+      .mockResolvedValueOnce([{ status: 'completed' }])
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(409)
+    await expect(response.json()).resolves.toEqual({
+      error: 'Execution cannot be cancelled while completed',
+    })
+    expect(mockWriteTerminalEvent).not.toHaveBeenCalled()
+    expect(mockCompletePausedCancellation).not.toHaveBeenCalled()
+    expect(mockClearPausedCancellationIntent).toHaveBeenCalledWith('ex-1', 'wf-1')
+  })
+
+  it('treats a concurrent cancellation as an idempotent success', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({ durablyRecorded: true, reason: 'recorded' })
+    dbChainMockFns.limit
+      .mockResolvedValueOnce([
+        { executionDeadlineAt: null, status: 'running', workspaceId: 'workspace-1' },
+      ])
+      .mockResolvedValueOnce([{ status: 'cancelled' }])
+    dbChainMockFns.returning.mockResolvedValueOnce([])
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toMatchObject({ success: true, reason: 'recorded' })
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
+    expect(mockWriteTerminalEvent).toHaveBeenCalledTimes(1)
+    expect(mockReleaseExecutionSlot).toHaveBeenCalledWith('ex-1')
+  })
+
+  it('updates execution log status in DB when durably recorded', async () => {
+    const mockWhere = vi.fn().mockResolvedValue(undefined)
+    const mockSet = vi.fn(() => ({ where: mockWhere }))
+    databaseMock.db.update.mockReturnValueOnce({ set: mockSet })
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: true,
+      reason: 'recorded',
+    })
+
+    await POST(makeRequest(), makeParams())
+
+    expect(databaseMock.db.update).toHaveBeenCalled()
+    expect(mockSet).toHaveBeenCalledWith({
+      status: 'cancelled',
+      endedAt: expect.any(Date),
+      totalDurationMs: expect.anything(),
+      executionDeadlineAt: null,
+    })
+  })
+
+  it('updates execution log status in DB when locally aborted', async () => {
+    const mockWhere = vi.fn().mockResolvedValue(undefined)
+    const mockSet = vi.fn(() => ({ where: mockWhere }))
+    databaseMock.db.update.mockReturnValueOnce({ set: mockSet })
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: false,
+      reason: 'redis_unavailable',
+    })
+    mockAbortManualExecution.mockReturnValue(true)
+
+    await POST(makeRequest(), makeParams())
+
+    expect(databaseMock.db.update).toHaveBeenCalled()
+    expect(mockSet).toHaveBeenCalledWith({
+      status: 'cancelled',
+      endedAt: expect.any(Date),
+      totalDurationMs: expect.anything(),
+      executionDeadlineAt: null,
+    })
+  })
+
+  it('claims the execution log before finalizing a paused cancellation', async () => {
+    mockStagePausedCancellation.mockResolvedValue({ kind: 'idle' })
+
+    await POST(makeRequest(), makeParams())
+
+    expect(databaseMock.db.update).toHaveBeenCalled()
+  })
+
+  it('does not confirm cancellation until the terminal database update succeeds', async () => {
+    mockMarkExecutionCancelled.mockResolvedValue({
+      durablyRecorded: true,
+      reason: 'recorded',
+    })
+    databaseMock.db.update.mockReturnValueOnce({
+      set: vi.fn(() => ({
+        where: vi.fn(() => {
+          throw new Error('DB connection failed')
+        }),
+      })),
+    })
+
+    const response = await POST(makeRequest(), makeParams())
+
+    expect(response.status).toBe(200)
+    const data = await response.json()
+    expect(data).toMatchObject({
+      success: false,
+      reason: 'cancellation_not_finalized',
+    })
+    expect(mockClearExecutionCancellation).not.toHaveBeenCalled()
   })
 })
