+++ apps/sim/lib/workflows/api/index.ts
@@ -2,6 +2,7 @@ export {
   createInternalWorkflowErrorPolicy,
   internalWorkflowErrorPolicies,
   internalWorkflowReadAuth,
+  internalWorkflowSessionOrApiKeyAuth,
   internalWorkflowSessionOrExecutorAuth,
   v2WorkflowErrorPolicies,
   WORKFLOW_NOT_FOUND_MESSAGE,
