+++ apps/sim/lib/workflows/application/cancel-run.ts
@@ -4,6 +4,7 @@ import {
   cancelWorkflowExecution,
   WorkflowExecutionNotFoundError,
 } from '@/lib/execution/cancel-workflow-execution'
+import { captureServerEvent } from '@/lib/posthog/server'
 import { defineAuthorizedWorkflowUseCase } from '@/lib/workflows/application/authorized-workflow-use-case'
 import { resolveActiveWorkflowRunApplicationContext } from '@/lib/workflows/application/context'
 import { workflowOperations } from '@/lib/workflows/application/operations'
@@ -11,6 +12,8 @@ import { workflowOperations } from '@/lib/workflows/application/operations'
 export interface CancelWorkflowRunInput {
   workflowId: string
   runId: string
+  assertedWorkspaceId?: string
+  abortSignal?: AbortSignal
 }
 
 export const cancelWorkflowRun = defineAuthorizedWorkflowUseCase({
@@ -19,8 +22,9 @@ export const cancelWorkflowRun = defineAuthorizedWorkflowUseCase({
     resolveActiveWorkflowRunApplicationContext({
       runId: input.runId,
       assertedWorkflowId: input.workflowId,
+      assertedWorkspaceId: input.assertedWorkspaceId,
     }),
-  async execute({ principal, context }) {
+  async execute({ principal, context, input }) {
     const attribution = resolvePrincipalAttribution(principal, {
       workspaceBillingOwnerUserId: context.billedAccountUserId,
     })
@@ -28,9 +32,9 @@ export const cancelWorkflowRun = defineAuthorizedWorkflowUseCase({
       const result = await cancelWorkflowExecution({
         executionId: context.runId,
         workflowId: context.workflowId,
-        userId: attribution.attributedUserId,
+        attributedUserId: attribution.attributedUserId,
         workspaceId: context.workspaceId,
-        captureAnalytics: false,
+        abortSignal: input.abortSignal,
       })
       return { ...result, workflowId: context.workflowId, workspaceId: context.workspaceId }
     } catch (error) {
@@ -39,5 +43,17 @@ export const cancelWorkflowRun = defineAuthorizedWorkflowUseCase({
       }
       throw error
     }
+  },
+  afterSuccess({ principal, context, result }) {
+    if (!result.success || result.reason === 'already_cancelled') return
+    const attribution = resolvePrincipalAttribution(principal, {
+      workspaceBillingOwnerUserId: context.billedAccountUserId,
+    })
+    captureServerEvent(
+      attribution.attributedUserId,
+      'workflow_execution_cancelled',
+      { workflow_id: context.workflowId, workspace_id: context.workspaceId },
+      { groups: { workspace: context.workspaceId } }
+    )
   },
 })
